@@ -4,10 +4,10 @@ build-backend = "flit_core.buildapi"
 
 [project]
 name="ansys-motorcad-core"
-version = "0.7.dev0"
+version = "0.9.dev0"
 description = "Pythonic interface to Ansys Motor-CAD."
 readme = "README.rst"
-requires-python = ">=3.8"
+requires-python = ">=3.9"
 license = {file = "LICENSE"}
 authors = [
     {name = "ANSYS, Inc.", email = "pyansys.core@ansys.com"},
@@ -21,35 +21,38 @@ classifiers = [
     "Topic :: Scientific/Engineering :: Physics",
     "License :: OSI Approved :: MIT License",
     "Operating System :: Microsoft :: Windows",
-    "Programming Language :: Python :: 3.7",
-    "Programming Language :: Python :: 3.8",
     "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
+    "Programming Language :: Python :: 3.11",
+    "Programming Language :: Python :: 3.12",
+    "Programming Language :: Python :: 3.13",
+    "Programming Language :: Python :: 3.14",
 ]
 dependencies = [
     "psutil >= 5.9.0",
     "requests >= 2.27.1",
     "packaging >= 21.3",
     "importlib-metadata>=4.0",
+    "openpyxl >=3.1.5",
 ]
 
 [project.optional-dependencies]
 tests = [
-    "pytest==8.3.3",
-    "pytest-cov==5.0.0",
+    "pytest==8.4.2",
+    "pytest-cov==7.0.0",
     "ansys.platform.instancemanagement>=1.0.2",
-    "matplotlib>=3.6.3"
+    "matplotlib>=3.6.3",
+    "scipy"
 ]
 doc = [
     "Sphinx==8.1.3",
-    "numpydoc==1.8.0",
-    "ansys-sphinx-theme==1.1.6",
+    "numpydoc==1.9.0",
+    "ansys-sphinx-theme==1.3.3",
     "sphinx-copybutton==0.5.2",
-    "sphinx-gallery==0.18.0",
+    "sphinx-gallery==0.19.0",
     "matplotlib",
     "scipy",
-    "pypandoc==1.14",
-    "bezier>=2023.7.28",
+    "pypandoc==1.16.2",
     "networkx"
 ]
 
