@@ -1,4 +1,4 @@
-# Copyright (C) 2022 - 2024 ANSYS, Inc. and/or its affiliates.
+# Copyright (C) 2022 - 2026 ANSYS, Inc. and/or its affiliates.
 # SPDX-License-Identifier: MIT
 #
 #
@@ -24,48 +24,225 @@
 from cmath import polar, rect
 from copy import deepcopy
 from enum import Enum
-from math import atan2, cos, degrees, inf, isclose, pi, radians, sin, sqrt
+from math import acos, atan2, comb, cos, degrees, fabs, floor, inf, isclose, radians, sin, sqrt
+import warnings
+from warnings import warn
+
+import ansys.motorcad.core
+from ansys.motorcad.core.geometry_extrusion import ExtrusionBlockList
 
 GEOM_TOLERANCE = 1e-6
 
 
-class Region(object):
-    """Python representation of Motor-CAD geometry region."""
+class RegionType(Enum):
+    """Provides an enumeration for storing Motor-CAD region types."""
 
-    def __init__(self, motorcad_instance=None):
-        """Create geometry region and set parameters to defaults."""
-        self.name = ""
-        self.material = "air"
-        self.colour = (0, 0, 0)
-        self.area = 0.0
-        self.centroid = Coordinate(0, 0)
-        self.region_coordinate = Coordinate(0, 0)
-        self.duplications = 1
-        self.entities = EntityList()
+    stator = "Stator"
+    rotor = "Rotor"
+    slot_area_stator = "Stator Slot Area"
+    slot_area_stator_deprecated = "Stator Slot"
+    slot_area_rotor = "Rotor Slot Area"
+    slot_split = "Split Slot"
+    stator_liner = "Stator Liner"
+    rotor_liner = "Rotor Liner"
+    wedge = "Wedge"
+    stator_duct = "Stator Duct"
+    housing_wj_wall = "Housing WJ Duct Wall"
+    housing = "Housing"
+    housing_magnetic = "Magnetic Housing"
+    stator_frame = "Stator Support Frame"
+    stator_impreg = "Stator Impreg"
+    impreg_gap = "Impreg Gap"
+    stator_copper = "Stator Copper"
+    stator_copper_ins = "Stator Copper Insulation"
+    stator_divider = "Stator Divider"
+    stator_slot_spacer = "Stator Slot Spacer"
+    stator_separator = "Stator slot separator"
+    coil_insulation = "Coil Insulation"
+    stator_air = "Stator Air"
+    endwinding = "End Winding"
+    rotor_hub = "Rotor hub"
+    rotor_air = "Rotor Air"
+    rotor_air_exc_liner = "Rotor Air (excluding liner area)"
+    rotor_pocket = "Rotor Pocket"
+    pole_spacer = "Pole Spacer"
+    rotor_slot = "Rotor Slot"
+    rotor_bar_end_ring = "Rotor Bar End Ring"
+    coil_separator = "Coil Separator"
+    damper_bar = "Damper Bar"
+    wedge_rotor = "Rotor Wedge"
+    rotor_divider = "Rotor Divider"
+    rotor_copper_ins = "Rotor Copper Insulation"
+    rotor_copper = "Rotor Copper"
+    rotor_impreg = "Rotor Impreg"
+    shaft = "Shaft"
+    axle = "Axle"
+    rotor_duct = "Rotor Duct"
+    magnet = "Magnet"
+    barrier = "Barrier"
+    mounting_base = "Base Mount"
+    mounting_plate = "Plate Mount"
+    endcap = "Endcap"
+    banding = "Banding"
+    sleeve = "Sleeve"
+    rotor_cover = "Rotor Cover"
+    slot_wj_insulation = "Slot Water Jacket Insulation"
+    slot_wj_wall = "Slot Water Jacket Wall"
+    slot_wj_duct = "Slot Water Jacket Duct"
+    slot_wj_duct_no_detail = "Slot Water Jacket Duct (no detail)"
+    cowling = "Cowling"
+    cowling_gril = "Cowling Grill"
+    cowling_grill_hole = "Cowling Grill Hole"
+    brush = "Brush"
+    bearings = "Bearings"
+    commutator = "Commutator"
+    airgap = "Airgap"
+    dxf_import = "DXF Import"
+    impreg_loss_lot_ac_loss = "Stator Proximity Loss Slot"
+    surrounding_air = "Surrounding air"
+    adaptive = "Adaptive Region"
+    no_type = "No type"
+
+
+RegionType.slot_area_stator_deprecated.__doc__ = "Only for use with Motor-CAD 2025.1 and earlier"
+
+
+class MagnetisationDirection(Enum):
+    """Provides an enumeration for storing Motor-CAD magnetisation direction."""
+
+    parallel = 0
+    radial = 1
+    function = 2
+
+
+class Region(object):
+    """Create geometry region.
+
+    Parameters
+    ----------
+    region_type: RegionType or str
+        Type of region. String must be a valid RegionType.
+    motorcad_instance: ansys.motorcad.core.MotorCAD
+        Motor-CAD instance currently connected
+    """
+
+    def __init__(self, region_type=RegionType.adaptive, motorcad_instance=None):
+        """Initialise Region."""
+        # handle the case where region_type is not a RegionType
+        if not isinstance(region_type, RegionType):
+            # if a string has been provided, try to set the region_type to this string
+            if isinstance(region_type, str):
+                try:
+                    # valid RegionType attributes will be set based on the string
+                    region_type = getattr(RegionType, region_type)
+                except Exception as e:
+                    raise Exception(
+                        f"{region_type} is not a valid RegionType. Please provide"
+                        f" a valid RegionType."
+                    )
+            # Try and catch case where user has added Motor-CAD instance without using named
+            # parameters
+            elif isinstance(region_type, ansys.motorcad.core.MotorCAD):
+                warnings.warn(
+                    "The first parameter of creating a new region has changed to region_type."
+                    " Please use named parameters ```Region(motorcad_instance=mc``` and add a"
+                    " region type"
+                )
+                motorcad_instance = region_type
+                region_type = region_type = RegionType.adaptive
+            # if any other type of object is provided, raise an exception.
+            else:
+                raise Exception(
+                    f"{region_type} is not a valid RegionType. Please provide a valid "
+                    f"RegionType."
+                )
+
+        # warn the user if they have not provided a specific region_type other than adaptive
+        elif region_type == RegionType.adaptive:
+            warnings.warn(
+                "It is strongly recommended to set a region_type when creating new regions."
+                " Creating new regions with no type will be deprecated in a future release"
+            )
+
+        self._name = ""
+        self._base_name = ""
+        self._material = "air"
+        self._colour = (255, 255, 255)
+        self._area = 0.0
+        self._centroid = Coordinate(0, 0)
+        self._region_coordinate = Coordinate(0, 0)
+        self._duplications = 1
+        self._entities = EntityList()
         self._parent_name = ""
         self._child_names = []
         self._motorcad_instance = motorcad_instance
-        self._region_type = RegionType.adaptive
-        self.mesh_length = 0
-        self._linked_region = None
+        self._region_type = region_type
+        self._mesh_length = 0
+        self._linked_region_names = []
+        self._extrusion_blocks = ExtrusionBlockList()
         self._singular = False
         self._lamination_type = ""
+
+        self._raw_region = dict()
 
     def __eq__(self, other):
         """Override the default equals implementation for Region."""
         return (
             isinstance(other, Region)
-            and self.name == other.name
-            and self.colour == other.colour
+            and self._name == other._name
             # and self.area == other.area ->
             # Already check entities - can't expect user to calculate area
             # and self.centroid == other.centroid ->
             # Centroid calculated from entities - can't expect user to calculate
             # and self.region_coordinate == other.region_coordinate ->
             # Region coordinate is an output, cannot guarantee will be same for identical regions
-            and self.duplications == other.duplications
-            and self.entities == other.entities
+            and self._duplications == other._duplications
+            and self._entities == other._entities
         )
+
+    def _get_new_object_of_type_self(self):
+        """Return self object."""
+        if self.region_type:
+            return type(self)(region_type=self.region_type)
+        else:
+            return type(self)()
+
+    def __copy__(self):
+        """Override default copy behaviour."""
+        copied_object = self._get_new_object_of_type_self()
+        copied_object.__dict__.update(self.__dict__)
+
+        # We don't want to copy raw json to a new region
+        # This could contain settings not visible or editable from PyMotorCAD object
+        copied_object._raw_region = dict()
+
+        return copied_object
+
+    def __deepcopy__(self, memo):
+        """Override default deepcopy behaviour."""
+
+        def override_close_motorcad_on_exit():
+            return False
+
+        copied_object = self._get_new_object_of_type_self()
+
+        memo[id(self)] = copied_object
+        for k, v in self.__dict__.items():
+            setattr(copied_object, k, deepcopy(v, memo))  # Copy over attributes
+
+        # We don't want to copy raw json to a new region
+        # This could contain settings not visible or editable from PyMotorCAD object
+        copied_object._raw_region = dict()
+
+        # Motor-CAD instance should not be duplicated.
+        # Don't want this getting closed if this region goes out of scope.
+        if copied_object._motorcad_instance is not None:
+            copied_object._motorcad_instance.connection._close_motorcad_on_exit = (
+                override_close_motorcad_on_exit
+            )
+            copied_object._motorcad_instance = self._motorcad_instance
+
+        return copied_object
 
     @classmethod
     def from_coordinate_list(cls):
@@ -83,7 +260,7 @@ class Region(object):
         entity : Line or Arc
             Line/arc entity class instance
         """
-        self.entities.append(entity)
+        self._entities.append(entity)
 
     def insert_entity(self, index, entity):
         """Insert entity to list of region entities at given index.
@@ -95,7 +272,7 @@ class Region(object):
         entity : Line or Arc
             Line/arc entity class instance
         """
-        self.entities.insert(index, entity)
+        self._entities.insert(index, entity)
 
     def insert_polyline(self, index, polyline):
         """Insert polyline at given index, polyline can be made up of line/arc entities.
@@ -110,6 +287,23 @@ class Region(object):
         for count, entity in enumerate(polyline):
             self.insert_entity(index + count, entity)
 
+    def is_closed(self):
+        """Check whether Region's entities are closed.
+
+        Returns
+        -------
+        Boolean
+            Whether Region is closed
+        """
+        if len(self.entities) > 0:
+            # Make sure all adjacent entities have a point in common
+            return all(
+                get_entities_have_common_coordinate(self.entities[i - 1], self.entities[i])
+                for i in range(0, len(self.entities))
+            )
+        else:
+            return False
+
     def remove_entity(self, entity_remove):
         """Remove the entity from the region.
 
@@ -118,15 +312,15 @@ class Region(object):
         entity_remove : Line or Arc
             Line/arc entity class instance
         """
-        for entity in self.entities:
+        for entity in self._entities:
             if (entity.start == entity_remove.start) & (entity.end == entity_remove.end):
                 if type(entity) == Line:
-                    self.entities.remove(entity)
+                    self._entities.remove(entity)
                 elif type(entity) == Arc:
                     if (entity.centre == entity_remove.centre) & (
                         entity.radius == entity_remove.radius
                     ):
-                        self.entities.remove(entity)
+                        self._entities.remove(entity)
 
     def replace(self, replacement_region):
         """Replace self with another region.
@@ -141,10 +335,90 @@ class Region(object):
             existing region.
         """
         # Remove existing entities from the region object
-        self.entities.clear()
+        self._entities.clear()
 
         # Set the region object entities to be the list of replacement region entities
-        self.entities = deepcopy(replacement_region.entities)
+        self._entities = deepcopy(replacement_region.entities)
+
+    def extend_entity(
+        self, entity_index, distance=None, fraction=None, factor=None, extend_from_end=True
+    ):
+        """Extend an Entity of a Region by a given distance.
+
+        The Entity will be extended by keeping
+        the start point fixed and shifting the end point unless extend_from_end is False. If
+        extend_from_end is False, the end point will be kept fixed and the start point will be
+        shifted instead. Tip: to extend from both ends, you can use the method twice.
+
+        Parameters
+        ----------
+        entity_index : int
+            Index of the Region Entity that is to be extended.
+        distance : float
+            Absolute distance to extend the Entity by. If negative, the Entity will be shortened.
+        fraction : float
+            Fractional distance to extend the Entity by. If negative, the Entity will be shortened.
+        factor : float
+            Factor to extend the Entity length to. The new Entity length will be the original
+            length multiplied by the factor. If less than 1, the Entity will be shortened.
+        extend_from_end : bool
+            If True, start point is unchanged and end point is shifted. If False, end point is
+            unchanged and start point is fixed.
+
+        """
+        if type(extend_from_end) == bool:
+            if extend_from_end:
+                point = self.entities[entity_index].end
+            else:
+                point = self.entities[entity_index].start
+        else:
+            raise TypeError(
+                f"The argument 'extend_from_end' must be a boolean type (True or " f"False)."
+            )
+
+        # if multiple optional arguments are provided, precedence is taken in the order: distance,
+        # fraction, factor:
+        if distance and fraction or distance and factor:
+            warn(f"More than one optional argument provided, using distance = {distance} mm.")
+            self.extend_entity(entity_index, distance=distance, extend_from_end=extend_from_end)
+            return
+
+        elif fraction and factor:
+            warn(f"More than one optional argument provided, using fraction = {fraction}.")
+            self.extend_entity(entity_index, fraction=fraction, extend_from_end=extend_from_end)
+            return
+
+        elif distance:
+            if distance <= -self.entities[entity_index].length:
+                raise ValueError(
+                    "Invalid distance provided. Cannot shorten entity by more than "
+                    "original length."
+                )
+            new_point = self.entities[entity_index].get_coordinate_from_distance(
+                point, distance=-distance
+            )
+        elif fraction:
+            if fraction <= -1:
+                raise ValueError(
+                    "Invalid fraction provided. Cannot shorten entity by more than "
+                    "original length."
+                )
+            new_point = self.entities[entity_index].get_coordinate_from_distance(
+                point, fraction=-fraction
+            )
+        elif factor:
+            if factor < 0:
+                factor = -factor
+            new_point = self.entities[entity_index].get_coordinate_from_distance(
+                point, fraction=1 - factor
+            )
+        else:
+            raise ValueError(f"Please provide either a distance, fraction or factor.")
+
+        if extend_from_end:
+            self.edit_point(self.entities[entity_index].end, new_point)
+        else:
+            self.edit_point(self.entities[entity_index].start, new_point)
 
     # method to receive region from Motor-CAD and create python object
     @classmethod
@@ -163,42 +437,54 @@ class Region(object):
 
         if is_magnet:
             new_region = RegionMagnet(motorcad_instance)
-            new_region._magnet_angle = json["magnet_angle"]
-            new_region._br_multiplier = json["magnet_magfactor"]
-            new_region._magnet_polarity = json["magnet_polarity"]
-            new_region._br_magnet = json["magnet_br_value"]
         else:
-            new_region = cls(motorcad_instance)
+            if has_region_type:
+                new_region = cls(
+                    motorcad_instance=motorcad_instance, region_type=RegionType(json["region_type"])
+                )
+            else:
+                new_region = cls(motorcad_instance=motorcad_instance)
 
-        if has_region_type:
-            new_region._region_type = RegionType(json["region_type"])
-
-        # self.Entities = json.Entities
-        new_region.name = json["name"]
-        new_region.material = json["material"]
-
-        new_region.colour = (json["colour"]["r"], json["colour"]["g"], json["colour"]["b"])
-        new_region.area = json["area"]
-
-        new_region.centroid = Coordinate(json["centroid"]["x"], json["centroid"]["y"])
-        new_region.region_coordinate = Coordinate(
-            json["region_coordinate"]["x"], json["region_coordinate"]["y"]
-        )
-        new_region.duplications = json["duplications"]
-        new_region.entities = _convert_entities_from_json(json["entities"])
-        new_region.parent_name = json["parent_name"]
-        new_region._child_names = json["child_names"]
-
-        if "mesh_length" in json:
-            new_region.mesh_length = json["mesh_length"]
-
-        if "singular" in json:
-            new_region._singular = json["singular"]
-
-        if "lamination_type" in json:
-            new_region._lamination_type = json["lamination_type"]
+        new_region._add_parameters_from_json(json)
 
         return new_region
+
+    def _add_parameters_from_json(self, json):
+        if "name_unique" in json:
+            self._name = json["name_unique"]
+        else:
+            self._name = json["name"]
+
+        self._base_name = json["name"]
+        self.material = json["material"]
+        self._colour = (json["colour"]["r"], json["colour"]["g"], json["colour"]["b"])
+        self._area = json["area"]
+
+        self._centroid = Coordinate(json["centroid"]["x"], json["centroid"]["y"])
+        self._region_coordinate = Coordinate(
+            json["region_coordinate"]["x"], json["region_coordinate"]["y"]
+        )
+        self._duplications = json["duplications"]
+        self._entities = _convert_entities_from_json(json["entities"])
+        self._parent_name = json["parent_name"]
+        self._child_names = json["child_names"]
+
+        if "mesh_length" in json:
+            self._mesh_length = json["mesh_length"]
+
+        if "singular" in json:
+            self._singular = json["singular"]
+
+        if "lamination_type" in json:
+            self._lamination_type = json["lamination_type"]
+
+        if "linked_regions" in json:
+            self._linked_region_names = json["linked_regions"]
+
+        if "extrusion_blocks" in json:
+            self._extrusion_blocks = ExtrusionBlockList._from_json(json["extrusion_blocks"])
+
+        self._raw_region = json
 
     # method to convert python object to send to Motor-CAD
     def _to_json(self):
@@ -209,61 +495,64 @@ class Region(object):
         dict
             Geometry region json representation
         """
-        if self._region_type == RegionType.adaptive:
-            lamination_type = self.lamination_type
-        else:
-            lamination_type = ""
+        # const for material component owner in geometry engine
+        COMPONENTOWNER_GEOMETRYENGINE = 2
 
-        region_dict = {
-            "name": self.name,
-            "material": self.material,
-            "colour": {"r": self.colour[0], "g": self.colour[1], "b": self.colour[2]},
-            "area": self.area,
-            "centroid": {"x": self.centroid.x, "y": self.centroid.y},
-            "region_coordinate": {"x": self.region_coordinate.x, "y": self.region_coordinate.y},
-            "duplications": self.duplications,
-            "entities": _convert_entities_to_json(self.entities),
-            "parent_name": self.parent_name,
-            "region_type": self._region_type.value,
-            "mesh_length": self.mesh_length,
-            "on_boundary": False if self._linked_region is None else True,
-            "singular": self._singular,
-            "lamination_type": lamination_type,
+        # Previous implementations had users only generally interact with the unique name,
+        # assigning it as the name attribute if possible. This behaviour is maintained for
+        # now, though it is a piece of information lost that future users may want control over
+        self._raw_region["name"] = self.name
+        if "name_unique" in self._raw_region:
+            self._raw_region["name_unique"] = self.name
+        self._raw_region["name_base"] = self._base_name
+
+        if (
+            (("material" in self._raw_region) and (self._raw_region["material"] != self._material))
+            or ("material" not in self._raw_region)
+            or (self._material == "")
+        ):
+            # material has changed from original material, update material component owner for
+            # geometry engine to create custom material component for this region
+            self._raw_region["material_weight_component_type"] = COMPONENTOWNER_GEOMETRYENGINE
+            # unable to assume material type from user material name, set to default of "Any"
+            self._raw_region["material_weight_material_type"] = "Any"
+
+        self._raw_region["material"] = self._material
+        self._raw_region["colour"] = {
+            "r": self._colour[0],
+            "g": self._colour[1],
+            "b": self._colour[2],
         }
+        self._raw_region["area"] = self._area
+        self._raw_region["centroid"] = {"x": self._centroid.x, "y": self._centroid.y}
+        self._raw_region["region_coordinate"] = {
+            "x": self._region_coordinate.x,
+            "y": self._region_coordinate.y,
+        }
+        self._raw_region["duplications"] = self._duplications
+        self._raw_region["entities"] = _convert_entities_to_json(self.entities)
+        self._raw_region["parent_name"] = self.parent_name
+        self._raw_region["region_type"] = self._region_type.value
+        self._raw_region["mesh_length"] = self.mesh_length
+        self._raw_region["linked_regions"] = self.linked_region_names
+        self._raw_region["singular"] = self._singular
+        self._raw_region["lamination_type"] = self._lamination_type
+        self._raw_region["extrusion_blocks"] = self._extrusion_blocks._to_json()
 
-        return region_dict
+        return self._raw_region
 
-    def is_closed(self):
-        """Check whether region entities create a closed region.
+    @property
+    def name(self):
+        """Name of Region."""
+        return self._name
 
-        Returns
-        -------
-        Boolean
-            Whether region is closed
-        """
-        if len(self.entities) > 0:
-            entity_first = self.entities[0]
-            entity_last = self.entities[-1]
-
-            is_closed = get_entities_have_common_coordinate(entity_first, entity_last)
-
-            for i in range(len(self.entities) - 1):
-                is_closed = get_entities_have_common_coordinate(
-                    self.entities[i], self.entities[i + 1]
-                )
-
-            return is_closed
-        else:
-            return False
+    @name.setter
+    def name(self, value):
+        self._name = value
 
     @property
     def parent_name(self):
-        """Get region parent name.
-
-        Returns
-        -------
-        string
-        """
+        """Get or set the region parent name."""
         return self._parent_name
 
     @parent_name.setter
@@ -273,16 +562,46 @@ class Region(object):
     @property
     def linked_region(self):
         """Get linked duplication/unite region."""
-        return self._linked_region
+        warn("linked_region property is deprecated. Use linked_regions array", DeprecationWarning)
+        return (
+            self.motorcad_instance.get_region(self._linked_region_names[0])
+            if len(self._linked_region_names) > 0
+            else None
+        )
 
     @linked_region.setter
     def linked_region(self, region):
-        self._linked_region = region
-        region._linked_region = self
+        warn(
+            "linked_region property is deprecated. Use linked_regions.append(region)",
+            DeprecationWarning,
+        )
+        self._linked_region_names.append(region.name)
+        region._linked_region_names.append(self.name)
+
+    @property
+    def linked_regions(self):
+        """Get linked region objects for duplication/unite operations."""
+        self._check_connection()
+        return [self.motorcad_instance.get_region(name) for name in self._linked_region_names]
+
+    @linked_regions.setter
+    def linked_regions(self, regions):
+        """Set linked regions for duplication/unite operations."""
+        self._linked_region_names = [region.name for region in regions]
+
+    @property
+    def linked_region_names(self):
+        """Get linked region names for duplication/unite operations."""
+        return self._linked_region_names
+
+    @linked_region_names.setter
+    def linked_region_names(self, names):
+        """Set linked region names for duplication/unite operations."""
+        self._linked_region_names = names
 
     @property
     def singular(self):
-        """Get linked duplication/unite region."""
+        """Get or set if region is singular."""
         return self._singular
 
     @singular.setter
@@ -290,8 +609,18 @@ class Region(object):
         self._singular = singular
 
     @property
+    def extrusion_blocks(self):
+        """Get extrusion blocks list.
+
+        Returns
+        -------
+        list of ExtrusionBlock
+        """
+        return self._extrusion_blocks
+
+    @property
     def child_names(self):
-        """Property for child names list.
+        """Get child names list.
 
         Returns
         -------
@@ -306,13 +635,17 @@ class Region(object):
 
         Returns
         -------
-        string
+        RegionType
         """
         return self._region_type
 
+    @region_type.setter
+    def region_type(self, region_type):
+        self._region_type = region_type
+
     @property
     def motorcad_instance(self):
-        """Get linked Motor-CAD instance."""
+        """Get or set the linked Motor-CAD instance."""
         return self._motorcad_instance
 
     @motorcad_instance.setter
@@ -336,7 +669,7 @@ class Region(object):
 
     @property
     def parent(self):
-        """Return parent region from Motor-CAD.
+        """Get or set parent region from Motor-CAD.
 
         Returns
         -------
@@ -352,7 +685,7 @@ class Region(object):
 
     @property
     def lamination_type(self):
-        """Return lamination type of region from Motor-CAD.
+        """Get or set lamination type of region from Motor-CAD.
 
         Returns
         -------
@@ -362,12 +695,81 @@ class Region(object):
 
     @lamination_type.setter
     def lamination_type(self, lamination_type):
-        if self.region_type == RegionType.adaptive:
-            self._lamination_type = lamination_type
-        else:
-            raise Exception(
-                "It is currently only possible to set lamination type for adaptive regions"
-            )
+        self._lamination_type = lamination_type
+
+    @property
+    def name(self):
+        """Get or set region name."""
+        return self._name
+
+    @name.setter
+    def name(self, name):
+        self._name = name
+
+    @property
+    def material(self):
+        """Get or set region material name."""
+        return self._material
+
+    @material.setter
+    def material(self, material):
+        self._material = material
+
+    @property
+    def colour(self):
+        """Get or set region colour."""
+        return self._colour
+
+    @colour.setter
+    def colour(self, colour):
+        self._colour = colour
+
+    @property
+    def duplications(self):
+        """Get or set number of region duplications for the full machine."""
+        return self._duplications
+
+    @duplications.setter
+    def duplications(self, duplications):
+        self._duplications = duplications
+
+    @property
+    def entities(self):
+        """Get or set the list of entities in the region."""
+        return self._entities
+
+    @entities.setter
+    def entities(self, entities):
+        self._entities = entities
+
+    @property
+    def mesh_length(self):
+        """Get or set the mesh length to use, or 0 for default."""
+        return self._mesh_length
+
+    @mesh_length.setter
+    def mesh_length(self, mesh_length):
+        self._mesh_length = mesh_length
+
+    @property
+    def area(self):
+        """Get the region area."""
+        return self._area
+
+    @property
+    def centroid(self):
+        """Get the region centroid."""
+        return self._centroid
+
+    @property
+    def region_coordinate(self):
+        """Get the reference coordinate within the region."""
+        return self._region_coordinate
+
+    @property
+    def duplication_angle(self):
+        """Get linked Motor-CAD instance."""
+        return 360 / self.duplications
 
     def subtract(self, region):
         """Subtract region from self, returning any additional regions.
@@ -505,15 +907,15 @@ class Region(object):
         """
         if isinstance(mirror_line, Line):
             region = deepcopy(self)
-            region.entities.clear()
-            region.centroid = self.centroid.mirror(mirror_line)
-            region.region_coordinate = self.region_coordinate.mirror(mirror_line)
+            region._entities.clear()
+            region._centroid = self._centroid.mirror(mirror_line)
+            region._region_coordinate = self._region_coordinate.mirror(mirror_line)
             region._child_names = []
 
             if unique_name:
-                region.name = region.name + "_mirrored"
+                region._name = region._name + "_mirrored"
 
-            for entity in self.entities:
+            for entity in self._entities:
                 region.add_entity(entity.mirror(mirror_line))
 
             return region
@@ -530,7 +932,7 @@ class Region(object):
         angle : float
             Angle of rotation in degrees. Anticlockwise direction is positive.
         """
-        for entity in self.entities:
+        for entity in self._entities:
             entity.rotate(centre_point, angle)
 
     def translate(self, x, y):
@@ -543,7 +945,7 @@ class Region(object):
         y : float
             y distance.
         """
-        for entity in self.entities:
+        for entity in self._entities:
             entity.translate(x, y)
 
     def update(self, region):
@@ -554,26 +956,26 @@ class Region(object):
         region : ansys.motorcad.core.geometry.Region
             Motor-CAD region object
         """
-        self.name = region.name
-        self.material = region.material
-        self.colour = region.colour
-        self.area = region.area
-        self.centroid = deepcopy(region.centroid)
-        self.region_coordinate = deepcopy(region.region_coordinate)
-        self.duplications = region.duplications
-        self.entities = deepcopy(region.entities)
-        self.parent_name = region.parent_name
-        self._child_names = region.child_names
+        self._name = region._name
+        self._material = region._material
+        self._colour = region._colour
+        self._area = region._area
+        self._centroid = deepcopy(region._centroid)
+        self._region_coordinate = deepcopy(region._region_coordinate)
+        self._duplications = region._duplications
+        self._entities = deepcopy(region._entities)
+        self._parent_name = region._parent_name
+        self._child_names = region._child_names
         self._motorcad_instance = region._motorcad_instance
 
     def _check_connection(self):
         """Check mc connection for region."""
-        if self.motorcad_instance is None:
+        if self._motorcad_instance is None:
             raise Exception(
                 "A Motor-CAD connection is required for this function"
                 + ", please set self.motorcad_instance to a valid Motor-CAD instance"
             )
-        if self.motorcad_instance.connection._wait_for_response(1) is False:
+        if self._motorcad_instance.connection._wait_for_response(1) is False:
             raise Exception(
                 "Unable to connect to Motor-CAD using self.motorcad_instance,"
                 + ", please set self.motorcad_instance to a valid Motor-CAD instance"
@@ -587,7 +989,7 @@ class Region(object):
         -------
         List of Coordinate
         """
-        return self.entities.points
+        return self._entities.points
 
     def add_point(self, point):
         """Add a new point into region on an existing Line/Arc.
@@ -600,7 +1002,7 @@ class Region(object):
         point : Coordinate
             Coordinate at which to add new point
         """
-        for pos, entity in enumerate(self.entities):
+        for pos, entity in enumerate(self._entities):
             if entity.coordinate_on_entity(point):
                 if isinstance(entity, Line):
                     new_entity_1 = Line(entity.start, point)
@@ -611,9 +1013,9 @@ class Region(object):
                 else:
                     raise Exception("Entity type is not Arc or Line")
 
-                self.entities.pop(pos)
-                self.entities.insert(pos, new_entity_1)
-                self.entities.insert(pos + 1, new_entity_2)
+                self._entities.pop(pos)
+                self._entities.insert(pos, new_entity_1)
+                self._entities.insert(pos + 1, new_entity_2)
                 break
 
         else:
@@ -629,7 +1031,7 @@ class Region(object):
         new_coordinates : Coordinate
             Position to move the point to
         """
-        for entity in self.entities:
+        for entity in self._entities:
             if entity.start == old_coordinates:
                 entity.start = deepcopy(new_coordinates)
             if entity.end == old_coordinates:
@@ -638,87 +1040,337 @@ class Region(object):
                 # Check Arc is still valid
                 _ = entity.centre
 
-    def round_corner(self, corner_coordinate, radius):
+    def _round_corner(self, corner_coordinate, radius, distance_limit):
         """Round the corner of a region.
+
+        The corner coordinates must already exist on two entities belonging to the region.
+        The two entities adjacent to the corner are shortened, and an arc is created between
+        them.
 
         Parameters
         ----------
         corner_coordinate : ansys.motorcad.core.geometry.Coordinate
-            Coordinate of the corner to be rounded.
+            Coordinate of the corner to round.
         radius : float
-            Radius by which the corner will be rounded.
+            Radius to round the corner by.
+        distance_limit : float
+            Maximum distance that the adjacent entities can be shortened by.
         """
-        adj_entities = []
-        entity_indicies = []
-        coordinates = []
-        # Get entities at corner
-        for index in range(len(self.entities)):
-            i = self.entities[index]
-            if i.coordinate_on_entity(corner_coordinate):
-                # coordinates.append(i.get_coordinate_from_distance(corner_coordinate, distance))
-                adj_entities.append(i)
-                entity_indicies.append(index)
-        if not adj_entities:
-            raise Exception(
+        # If radius is 0, do nothing
+        if radius == 0:
+            return
+
+        # Find adjacent entities. There should be 2 entities adjacent to the corner. Going
+        # anti-clockwise around the region, the entities before and after the corner will be
+        # adj_entity[0] and adj_entity[1] respectively.
+        adj_entities = [None, None]
+        adj_entity_indices = [None, None]
+        for index in range(len(self._entities)):
+            entity = self._entities[index]
+            if entity.end == corner_coordinate:
+                adj_entities[0] = entity
+                adj_entity_indices[0] = index
+            elif entity.start == corner_coordinate:
+                adj_entities[1] = entity
+                adj_entity_indices[1] = index
+
+        # If no adjacent entities are found, the point provided is not a corner
+        if not adj_entities[0] and not adj_entities[1]:
+            raise ValueError(
                 "Failed to find point on entity in region. "
                 "You must specify a corner in this region."
             )
-        if len(adj_entities) == 1:
-            raise Exception(
+        # If only one adjacent entity is found, the point provided is not a corner
+        if not adj_entities[0] or not adj_entities[1]:
+            raise ValueError(
                 "Point found on only one entity in region. "
                 "You must specify a corner in this region."
             )
-        # if first and last entities, swap the positions
-        if entity_indicies[0] == 0 and entity_indicies[1] == len(self.entities) - 1:
-            entity_indicies[0], entity_indicies[1] = entity_indicies[1], entity_indicies[0]
-            adj_entities[0], adj_entities[1] = adj_entities[1], adj_entities[0]
-        # Calculate distances for adjacent entities
-        if type(adj_entities[0]) == Line:
-            angle_1 = degrees(
-                atan2(
-                    adj_entities[0].start.y - adj_entities[0].end.y,
-                    adj_entities[0].start.x - adj_entities[0].end.x,
-                )
-            )
-        else:  # Arc
-            point_on_curve = adj_entities[0].get_coordinate_from_distance(corner_coordinate, 0.01)
-            angle_1 = degrees(
-                atan2(
-                    point_on_curve.y - adj_entities[0].end.y,
-                    point_on_curve.x - adj_entities[0].end.x,
-                )
-            )
-        if type(adj_entities[1]) == Line:
-            angle_2 = degrees(
-                atan2(
-                    adj_entities[1].end.y - adj_entities[1].start.y,
-                    adj_entities[1].end.x - adj_entities[1].start.x,
-                )
-            )
-        else:  # Arc
-            point_on_curve = adj_entities[1].get_coordinate_from_distance(corner_coordinate, 0.01)
-            angle_2 = degrees(
-                atan2(
-                    point_on_curve.y - adj_entities[1].start.y,
-                    point_on_curve.x - adj_entities[1].start.x,
-                )
-            )
 
-        corner_angle = abs(angle_1 - angle_2)
-        arc_angle = 90 + (90 - corner_angle)
-        half_chord = radius * sin(radians(arc_angle) / 2)
-        distance = abs(half_chord / (sin(radians(corner_angle) / 2)))
+        # If we have arc rounding, we need to find the angle at the intersection of the arc and the
+        # rounding arc. We don't know this position in advance, so iterate up to 100 times to find
+        # the correct distance.
+        distance = 0
+        converged = False
+        for iteration in range(100):
+            # get the angles of the adjacent entities. For a line, this is a property of the entity
+            # object. For an arc, approximate the arc by a straight line from the arc start or end
+            # (whichever is the corner coordinate) to a point 0.0001 mm along the arc.
+            adj_entity_angles = []
+            for entity in adj_entities:
+                if isinstance(entity, Arc):
+                    if corner_coordinate == entity.start:
+                        point_on_arc1 = entity.get_coordinate_from_distance(entity.start, distance)
+                        point_on_arc2 = entity.get_coordinate_from_distance(
+                            entity.start, distance + 0.0001
+                        )
+                    else:
+                        point_on_arc1 = entity.get_coordinate_from_distance(
+                            entity.end, distance + 0.0001
+                        )
+                        point_on_arc2 = entity.get_coordinate_from_distance(entity.end, distance)
+                    line_on_arc = Line(point_on_arc1, point_on_arc2)
+                    adj_entity_angles.append(line_on_arc.angle)
+                else:
+                    adj_entity_angles.append(entity.angle)
 
-        for index in range(len(adj_entities)):
-            j = adj_entities[index]
-            coordinates.append(j.get_coordinate_from_distance(corner_coordinate, distance))
-            if j.start == corner_coordinate:
-                j.start = coordinates[index]
-            elif j.end == corner_coordinate:
-                j.end = coordinates[index]
+            # calculate the internal angle of the corner.
+            corner_internal_angle = 180 + (adj_entity_angles[0] - adj_entity_angles[1])
+            # If this is more than 360, subtract 360.
+            if corner_internal_angle > 360:
+                corner_internal_angle = corner_internal_angle - 360
+            # If it is less than zero, add 360.
+            elif corner_internal_angle < 0:
+                corner_internal_angle = corner_internal_angle + 360
 
-        corner_arc = Arc(coordinates[0], coordinates[1], radius=radius)
-        self.insert_entity(entity_indicies[0] + 1, corner_arc)
+            # calculate the arc angle
+            corner_arc_angle = 180 - corner_internal_angle
+
+            # If the arc angle is zero, the point provided is not a corner.
+            if (
+                isclose(corner_arc_angle, 0, abs_tol=1e-3)
+                or isclose(corner_arc_angle, 360, abs_tol=1e-3)
+                or isclose(corner_arc_angle, -360, abs_tol=1e-3)
+            ):
+                return
+
+            # Calculate distances by which the adjacent entities are shortened
+            previous_distance = distance
+            half_chord = radius * sin(radians(corner_arc_angle) / 2)
+            distance = abs(half_chord / (sin(radians(corner_internal_angle) / 2)))
+
+            # Check if distance has converged, or if iterative convergence not needed
+            if (isinstance(adj_entities[0], Line) and isinstance(adj_entities[1], Line)) or isclose(
+                previous_distance, distance, abs_tol=1e-3
+            ):
+                converged = True
+                break
+
+        # Raise assertion if not converged, as radius probably not valid
+        if converged == False:
+            raise ValueError("Cannot find intersection. Check if radius is too large")
+
+        # check that the  distances by which the adjacent entities are shortened are less than
+        # the distance_limit.
+        if distance_limit < distance:
+            raise ValueError(
+                "Corner radius is too large for these entities. "
+                "You must specify a smaller radius."
+            )
+        # get and set the new start and end coordinates for the adjacent entities
+        adj_entities[0].end = adj_entities[0].get_coordinate_from_distance(
+            corner_coordinate, distance
+        )
+        adj_entities[1].start = adj_entities[1].get_coordinate_from_distance(
+            corner_coordinate, distance
+        )
+
+        # if the internal angle of the corner is more than 180, a negative radius must be applied
+        if corner_internal_angle > 180:
+            e = -1
+        else:
+            e = 1
+
+        # create the round corner arc and insert at the index after the first adjacent entity.
+        corner_arc = Arc(adj_entities[0].end, adj_entities[1].start, radius=e * radius)
+        self.insert_entity(adj_entity_indices[0] + 1, corner_arc)
+
+    def round_corner(self, corner_coordinate, radius, maximise=True):
+        """Round the corner of a region.
+
+        The corner coordinates must already exist on two entities belonging to the region.
+        The two entities adjacent to the corner are shortened, and an arc is created between
+        them.
+
+        Parameters
+        ----------
+        corner_coordinate : ansys.motorcad.core.geometry.Coordinate
+            Coordinate of the corner to round.
+        radius : float
+            Radius to round the corner by.
+        maximise : bool
+            Whether to maximise the possible radius if the radius provided is too large.
+        """
+        # get the lengths of the original adjacent entities before any corner rounding
+        adj_entity_lengths = []
+        for entity in self._entities:
+            if entity.end == corner_coordinate or entity.start == corner_coordinate:
+                adj_entity_lengths.append(entity.length)
+        # find the limit for how much an adjacent entity may be shortened by:
+        distance_limit = 10000
+        for length in adj_entity_lengths:
+            if length < distance_limit:
+                distance_limit = length
+
+        # if the corner radius is too large, an exception will be raised
+        if not maximise:
+            # round the corner
+            self._round_corner(corner_coordinate, radius, distance_limit)
+        # if the corner radius is too large, maximise the corner radius that is short enough for
+        # the adjacent entities. The adjacent entities can only be shortened to half the original
+        # entity length before any corner rounding
+        else:
+            try:
+                # try to round the corner with the specified radius
+                self._round_corner(corner_coordinate, radius, distance_limit)
+            except ValueError as e:
+                if "Corner radius is too large for these entities" in str(e):
+                    new_corner_radius = round(radius, 1)
+                    # iterate 100 times to find a maximum suitable corner radius
+                    for iteration in range(100):
+                        # subtract 0.1 mm from the previous corner radius that was tried
+                        new_corner_radius -= 0.1
+                        try:
+                            # try to round the corner with the new shorter radius
+                            self._round_corner(corner_coordinate, new_corner_radius, distance_limit)
+                            break
+                        except ValueError as e:
+                            if "Corner radius is too large for these entities" in str(e):
+                                # try 100 iterations
+                                if iteration < 99:
+                                    pass
+                                # if the radius is still too large on the final iteration, raise
+                                # the exception
+                                else:
+                                    raise e
+                # if any other exception is raised by the corner rounding attempt, raise the
+                # exception
+                else:
+                    raise e
+
+    def round_corners(self, corner_coordinates, radius, maximise=True):
+        """Round multiple corners of a region.
+
+        Each corner coordinate must already exist on two entities belonging to the region.
+        The two entities adjacent to each corner are shortened, and an arc is created
+        between them.
+
+        Parameters
+        ----------
+        corner_coordinates : list of ansys.motorcad.core.geometry.Coordinate
+            List of coordinates of the corners to round.
+        radius : float
+            Radius to round the corners by.
+        maximise : bool
+            Whether to maximise the possible radius of each corner if the radius provided is too
+            large.
+        """
+        # if the corner radius is too large, an exception will be raised
+        if not maximise:
+            # get the original entities before any corner rounding has been done
+            entities_orig = deepcopy(self._entities)
+            # apply the rounding to each corner in turn
+            for corner in corner_coordinates:
+                # get the lengths of the original adjacent entities before any corner rounding
+                adj_entity_lengths = []
+                for index in range(len(entities_orig)):
+                    entity = entities_orig[index]
+                    if entity.end == corner or entity.start == corner:
+                        adj_entity_lengths.append(entity.length)
+
+                # find the distance limit that the adjacent entities can be shortened by
+                distance_limit = 10000
+                for index in range(len(adj_entity_lengths)):
+                    if adj_entity_lengths[index] < distance_limit:
+                        distance_limit = adj_entity_lengths[index]
+
+                # round the corner
+                self._round_corner(corner, radius, distance_limit / 2)
+        # if the corner radius is too large, maximise the corner radius that is short enough for
+        # the adjacent entities. The adjacent entities can only be shortened to half the original
+        # entity length before any corner rounding
+        else:
+            # get the original entities before any corner rounding has been done
+            entities_orig = deepcopy(self._entities)
+            # apply the rounding to each corner in turn
+            for corner in corner_coordinates:
+                # get the lengths of the original adjacent entities before any corner rounding
+                adj_entity_lengths = []
+                for index in range(len(entities_orig)):
+                    entity = entities_orig[index]
+                    if entity.end == corner or entity.start == corner:
+                        adj_entity_lengths.append(entity.length)
+                # find the distance limit that the adjacent entities can be shortened by
+                distance_limit = 10000
+                for index in range(len(adj_entity_lengths)):
+                    if adj_entity_lengths[index] < distance_limit:
+                        distance_limit = adj_entity_lengths[index]
+                try:
+                    # try to round the corner with the specified radius
+                    self._round_corner(corner, radius, distance_limit / 2)
+                except ValueError as e:
+                    if "Corner radius is too large for these entities" in str(e):
+                        new_corner_radius = round(radius, 1)
+                        # iterate 100 times to find a maximum suitable corner radius
+                        for iteration in range(100):
+                            # subtract 0.1 mm from the previous corner radius that was tried
+                            new_corner_radius -= 0.1
+                            try:
+                                # try to round the corner with the new shorter radius
+                                self._round_corner(corner, new_corner_radius, distance_limit / 2)
+                                break
+                            except ValueError as e:
+                                if "Corner radius is too large for these entities" in str(e):
+                                    # try 100 iterations
+                                    if iteration < 99:
+                                        pass
+                                    # if the radius is still too large on the final iteration, raise
+                                    # the exception
+                                    else:
+                                        raise e
+                    # if any other exception is raised by the corner rounding attempt, raise the
+                    # exception
+                    else:
+                        raise e
+
+    def limit_arc_chord(self, max_chord_height):
+        """Limit the chord height for arcs in a region.
+
+        Subdivide arcs if required to ensure the arc's chord height (the distance between the arc
+        midpoint and the midpoint of a line between the start and end) is lower than the specified
+        value. This can be used to force a fine FEA mesh around entities with high curvature.
+
+        Parameters:
+            max_chord_height: float
+                The maximum chord height allowed.
+        """
+        new_entity_list = []
+        for entity in self.entities:
+            if (
+                isinstance(entity, Arc)
+                and (entity.radius != 0)
+                and (fabs(entity.radius * 2) > max_chord_height)
+            ):
+                # Find maximum arc angle so the chord height is equal to the required maximum error
+                max_angle = degrees(2 * acos(1 - max_chord_height / fabs(entity.radius)))
+
+                # Find how many segments are needed to achieve this
+                if max_angle > 0:
+                    segmentation = floor(entity.total_angle / max_angle)
+                else:
+                    segmentation = 1
+
+                # If required, split the arc into segments
+                if segmentation > 1:
+                    segment_start = entity.start
+                    for segment in range(1, segmentation + 1):
+                        segment_end = entity.get_coordinate_from_distance(
+                            entity.start, fraction=segment / segmentation
+                        )
+                        new_arc = Arc(segment_start, segment_end, entity.centre, entity.radius)
+                        # Add to the list
+                        new_entity_list.append(new_arc)
+                        # Ready for next segment
+                        segment_start = segment_end
+                else:
+                    # Arc already OK, just add to list as-is
+                    new_entity_list.append(entity)
+            else:
+                # Add to list unchanged
+                new_entity_list.append(entity)
+
+        # Update the entity list
+        self.entities = new_entity_list
 
     def find_entity_from_coordinates(self, coordinate_1, coordinate_2):
         """Search through region to find an entity with start and end coordinates.
@@ -734,7 +1386,7 @@ class Region(object):
         -------
         Line or Arc entity
         """
-        for entity in self.entities:
+        for entity in self._entities:
             if (coordinate_1 == entity.start) and (coordinate_2 == entity.end):
                 return entity
             elif (coordinate_1 == entity.end) and (coordinate_2 == entity.start):
@@ -744,16 +1396,37 @@ class Region(object):
 
 
 class RegionMagnet(Region):
-    """Provides the Python representation of a Motor-CAD magnet geometry region."""
+    """Create magnet geometry region.
+
+    Parameters
+    ----------
+    motorcad_instance: ansys.motorcad.core.MotorCAD
+        Motor-CAD instance currently connected
+    """
 
     def __init__(self, motorcad_instance=None):
-        """Initialise a ``RegionMagnet`` instance."""
-        super().__init__(motorcad_instance)
+        """Initialise Magnet Region."""
+        super().__init__(RegionType.magnet, motorcad_instance)
         self._magnet_angle = 0.0
-        self._br_multiplier = 0.0
+        self._br_multiplier = 1.0
         self._br_magnet = 0.0
         self._magnet_polarity = ""
-        self._region_type = RegionType.magnet
+        self.magnetisation_direction = MagnetisationDirection.parallel
+        self.magnetisation_function_amplitude = ""
+        self.magnetisation_function_angle = ""
+
+    def _add_parameters_from_json(self, json):
+        self._magnet_angle = json["magnet_angle"]
+        self._br_multiplier = json["magnet_magfactor"]
+        self._magnet_polarity = json["magnet_polarity"]
+        self._br_magnet = json["magnet_br_value"]
+
+        if "magnetisation_direction" in json:
+            self.magnetisation_direction = MagnetisationDirection(json["magnetisation_direction"])
+            self.magnetisation_function_amplitude = json["magnetisation_function_amplitude"]
+            self.magnetisation_function_angle = json["magnetisation_function_angle"]
+
+        super()._add_parameters_from_json(json)
 
     def _to_json(self):
         """Convert from a Python class to a JSON object.
@@ -767,6 +1440,10 @@ class RegionMagnet(Region):
 
         region_dict["magnet_magfactor"] = self._br_multiplier
         region_dict["magnet_angle"] = self._magnet_angle
+        region_dict["magnet_polarity"] = self._magnet_polarity
+        region_dict["magnetisation_direction"] = self.magnetisation_direction.value
+        region_dict["magnetisation_function_amplitude"] = self.magnetisation_function_amplitude
+        region_dict["magnetisation_function_angle"] = self.magnetisation_function_angle
 
         return region_dict
 
@@ -847,6 +1524,10 @@ class RegionMagnet(Region):
         string
         """
         return self._magnet_polarity
+
+    @magnet_polarity.setter
+    def magnet_polarity(self, value):
+        self._magnet_polarity = value
 
 
 class Coordinate(object):
@@ -962,6 +1643,20 @@ class Coordinate(object):
         self.x += x
         self.y += y
 
+    def distance(self, other):
+        """Find the distance to a coordinate.
+
+        Parameters
+        ----------
+        other: Coordinate
+            Find distance to this point.
+
+        Returns
+        -------
+        float
+        """
+        return abs(other - self)
+
     @classmethod
     def from_polar_coords(cls, radius, theta):
         """Create Coordinate from polar coordinates.
@@ -1050,6 +1745,114 @@ class Entity(object):
         self.start.translate(x, y)
         self.end.translate(x, y)
 
+    def get_intersection(self, entity):
+        """Get intersection Coordinate of entity with another entity.
+
+        Returns None if intersection not found.
+
+        Parameters
+        ----------
+        entity : ansys.motorcad.core.geometry.Line or ansys.motorcad.core.geometry.Arc
+
+        Returns
+        -------
+        ansys.motorcad.core.geometry.Coordinate or list of Coordinate or None
+        """
+        if isinstance(self, Line):
+            if isinstance(entity, Line):
+                points = self.get_line_intersection(entity)
+            elif isinstance(entity, Arc):
+                points = entity.get_line_intersection(self)
+            else:
+                raise Exception("Entity type is not Arc or Line")
+        else:
+            if isinstance(entity, Line):
+                points = self.get_line_intersection(entity)
+            elif isinstance(entity, Arc):
+                points = self.get_arc_intersection(entity)
+        if points:
+            intersections = []
+            if type(points) == list:
+                for point in points:
+                    if self.coordinate_on_entity(point):
+                        intersections.append(point)
+            else:
+                if self.coordinate_on_entity(points):
+                    intersections.append(points)
+            if intersections:
+                return intersections
+            else:
+                return None
+        else:
+            return None
+
+    def extend(self, distance=None, fraction=None, factor=None, extend_from_end=True):
+        """Extend an Entity object by a given distance.
+
+        The Entity will be extended by keeping the
+        start point fixed and shifting the end point unless extend_from_end is False. If
+        extend_from_end is False, the end point will be kept fixed and the start point will be
+        shifted instead. Tip: To extend from both ends, you can use the method twice.
+
+        Parameters
+        ----------
+        distance : float
+            Absolute distance to extend the Entity by. If negative, the Entity will be shortened.
+        fraction : float
+            Fractional distance to extend the Entity by. If negative, the Entity will be shortened.
+        factor : float
+            Factor to extend the Entity length to. The new Entity length will be the original
+            length multiplied by the factor. If less than 1, the Entity will be shortened.
+        extend_from_end : bool
+            If True, start point is unchanged and end point is shifted. If False, end point is
+            unchanged and start point is fixed.
+
+        """
+        if type(extend_from_end) == bool:
+            if extend_from_end:
+                point = self.end
+            else:
+                point = self.start
+        else:
+            raise TypeError(
+                f"The argument 'extend_from_end' must be a boolean type (True or False)."
+            )
+
+        # if multiple optional arguments are provided, precedence is taken in the order: distance,
+        # fraction, factor:
+        if distance and fraction or distance and factor:
+            warn(f"More than one optional argument provided, using distance = {distance} mm.")
+            self.extend(distance=distance, extend_from_end=extend_from_end)
+            return
+        elif fraction and factor:
+            warn(f"More than one optional argument provided, using fraction = {fraction}.")
+            self.extend(fraction=fraction, extend_from_end=extend_from_end)
+            return
+        elif distance:
+            if distance <= -self.length:
+                raise ValueError(
+                    "Invalid distance provided. Cannot shorten entity by more than "
+                    "original length."
+                )
+            new_point = self.get_coordinate_from_distance(point, distance=-distance)
+        elif fraction:
+            if fraction <= -1:
+                raise ValueError(
+                    "Invalid fraction provided. Cannot shorten entity by more than "
+                    "original length."
+                )
+            new_point = self.get_coordinate_from_distance(point, fraction=-fraction)
+        elif factor:
+            if factor < 0:
+                factor = -factor
+            new_point = self.get_coordinate_from_distance(point, fraction=1 - factor)
+        else:
+            raise ValueError(f"Please provide either a distance, fraction or factor.")
+        if extend_from_end:
+            self.end = new_point
+        else:
+            self.start = new_point
+
 
 class Line(Entity):
     """Python representation of Motor-CAD line entity based upon start and end coordinates.
@@ -1070,6 +1873,22 @@ class Line(Entity):
     def __eq__(self, other):
         """Override the default equals implementation for Line."""
         return isinstance(other, Line) and self.start == other.start and self.end == other.end
+
+    def get_bounds(self):
+        """
+        Return relevant information about the maximum and minimum x, y, and radius.
+
+        Returns: (max_radius, max_x, min_x, max_y, min_y)
+        """
+        max_radius = max(
+            Coordinate.get_polar_coords_deg(self.start)[0],
+            Coordinate.get_polar_coords_deg(self.end)[0],
+            Coordinate.get_polar_coords_deg(self.midpoint)[0],
+        )
+        xs = (self.start.x, self.end.x, self.midpoint.x)
+        ys = (self.start.y, self.end.y, self.midpoint.y)
+
+        return (max_radius, max(xs), min(xs), max(ys), min(ys))
 
     @property
     def midpoint(self):
@@ -1148,36 +1967,37 @@ class Line(Entity):
         else:
             raise Exception("Line can only be mirrored about Line()")
 
-    def get_coordinate_from_percentage_distance(self, ref_coordinate, percentage):
-        """Get the coordinate at the percentage distance along the line from the reference.
+    def get_coordinate_from_percentage_distance(self, ref_coordinate, fraction):
+        """Get the coordinate at a fractional distance along the line from the reference coord.
+
+        .. note::
+           This method is deprecated. Use the :func:`Line.get_coordinate_from_distance`
+           method with the `fraction = ` or `percentage =` argument.
 
         Parameters
         ----------
         ref_coordinate : Coordinate
             Entity reference coordinate.
 
-        percentage : float
-            Percentage distance along Line.
+        fraction : float
+            Fractional distance along Line.
 
         Returns
         -------
         Coordinate
-            Coordinate at percentage distance along Line.
+            Coordinate at fractional distance along Line.
         """
-        if ref_coordinate == self.end:
-            coordinate_1 = self.end
-            coordinate_2 = self.start
-        else:
-            coordinate_1 = self.start
-            coordinate_2 = self.end
+        warn(
+            "get_coordinate_from_percentage_distance() WILL BE DEPRECATED SOON - "
+            "USE get_coordinate_from_distance instead with the `fraction = ` or `percentage = ` "
+            "optional argument",
+            DeprecationWarning,
+        )
+        return self.get_coordinate_from_distance(ref_coordinate, fraction=fraction)
 
-        t = (self.length * percentage) / self.length
-        x = ((1 - t) * coordinate_1.x) + (t * coordinate_2.x)
-        y = ((1 - t) * coordinate_1.y) + (t * coordinate_2.y)
-
-        return Coordinate(x, y)
-
-    def get_coordinate_from_distance(self, ref_coordinate, distance):
+    def get_coordinate_from_distance(
+        self, ref_coordinate, distance=None, fraction=None, percentage=None
+    ):
         """Get the coordinate at the specified distance along the line from the reference.
 
         Parameters
@@ -1185,22 +2005,54 @@ class Line(Entity):
         ref_coordinate : Coordinate
             Entity reference coordinate.
 
-        distance : float
+        distance : float, optional
             Distance along Line.
+
+        fraction : float, optional
+            Fractional distance along Line.
+
+        percentage : float, optional
+            Percentage distance along Line.
 
         Returns
         -------
         Coordinate
             Coordinate at distance along Line.
         """
+        if (distance is None) and (fraction is None) and (percentage is None):
+            raise Exception("You must provide either a distance, fraction or percentage.")
+
+        if (distance is not None) and (fraction is not None):
+            warn("Both distance and fraction provided. Using distance.", UserWarning)
+        if (distance is not None) and (percentage is not None):
+            warn("Both distance and percentage provided. Using distance.", UserWarning)
+
+        if distance is None:
+            if (fraction is not None) and (percentage is not None):
+                warn("Both fraction and percentage provided. Using fraction.", UserWarning)
+            if fraction is not None:
+                distance = self.length * fraction
+            elif percentage is not None:
+                distance = self.length * (percentage / 100)
+
         if ref_coordinate == self.end:
             coordinate_1 = self.end
             coordinate_2 = self.start
-        else:
+        elif ref_coordinate == self.start:
             coordinate_1 = self.start
             coordinate_2 = self.end
+        elif self.coordinate_on_entity(ref_coordinate):
+            if ref_coordinate.distance(self.start) <= ref_coordinate.distance(self.end):
+                # Ref coordinate closer to start, so go between ref and end
+                coordinate_1 = ref_coordinate
+                coordinate_2 = self.end
+            else:
+                coordinate_1 = ref_coordinate
+                coordinate_2 = self.start
+        else:
+            raise ValueError("Invalid ref_coordinate provided, it must be on the line.")
 
-        t = distance / self.length
+        t = distance / coordinate_1.distance(coordinate_2)
         x = ((1 - t) * coordinate_1.x) + (t * coordinate_2.x)
         y = ((1 - t) * coordinate_1.y) + (t * coordinate_2.y)
 
@@ -1279,6 +2131,32 @@ class Line(Entity):
             # Lines don't intersect
             return None
 
+    def get_arc_intersection(self, arc):
+        """Get intersection Coordinates of line with an arc.
+
+        Returns None if intersection not found.
+
+        Parameters
+        ----------
+        arc : ansys.motorcad.core.geometry.Arc
+
+        Returns
+        -------
+        ansys.motorcad.core.geometry.Coordinate or list of Coordinate or None
+        """
+        return arc.get_line_intersection(self)
+
+    def get_coordinate_distance(self, coordinate):
+        """Get distance of line with another coordinate."""
+        normal_angle = self.angle - 90
+        defining_point = Coordinate.from_polar_coords(1, normal_angle)
+        normal = Line(Coordinate(0, 0), defining_point)
+        normal.translate(coordinate.x, coordinate.y)
+        nearest_point = self.get_line_intersection(normal)
+        if nearest_point is None:
+            return None
+        return sqrt((coordinate.x - nearest_point.x) ** 2 + (coordinate.y - nearest_point.y) ** 2)
+
 
 class _BaseArc(Entity):
     """Internal class to allow creation of Arcs."""
@@ -1312,27 +2190,37 @@ class _BaseArc(Entity):
         x_shift, y_shift = rt_to_xy(abs(self.radius), angle)
         return Coordinate(self.centre.x + x_shift, self.centre.y + y_shift)
 
-    def get_coordinate_from_percentage_distance(self, ref_coordinate, percentage):
-        """Get the coordinate at the percentage distance along the arc from the reference coord.
+    def get_coordinate_from_percentage_distance(self, ref_coordinate, fraction):
+        """Get the coordinate at a fractional distance along the arc from the reference coord.
+
+        .. note::
+           This method is deprecated. Use the :func:`Arc.get_coordinate_from_distance`
+           method with the `fraction = ` or `percentage =` argument.
 
         Parameters
         ----------
         ref_coordinate : Coordinate
             Entity reference coordinate.
 
-        percentage : float
-            Percentage distance along Arc.
+        fraction : float
+            Fractional distance along Arc.
 
         Returns
         -------
         Coordinate
-            Coordinate at percentage distance along Arc.
+            Coordinate at fractional distance along Arc.
         """
-        length = self.length * percentage
+        warn(
+            "get_coordinate_from_percentage_distance() WILL BE DEPRECATED SOON - "
+            "USE get_coordinate_from_distance instead with the `fraction = ` or `percentage = ` "
+            "optional argument",
+            DeprecationWarning,
+        )
+        return self.get_coordinate_from_distance(ref_coordinate, fraction=fraction)
 
-        return self.get_coordinate_from_distance(ref_coordinate, length)
-
-    def get_coordinate_from_distance(self, ref_coordinate, distance):
+    def get_coordinate_from_distance(
+        self, ref_coordinate, distance=None, fraction=None, percentage=None
+    ):
         """Get the coordinate at the specified distance along the arc from the reference coordinate.
 
         Parameters
@@ -1340,14 +2228,36 @@ class _BaseArc(Entity):
         ref_coordinate : Coordinate
            Entity reference coordinate.
 
-        distance : float
+        distance : float, optional
             Distance along arc.
+
+        fraction : float, optional
+            Fractional distance along Arc.
+
+        percentage : float, optional
+            Percentage distance along Arc.
 
         Returns
         -------
         Coordinate
             Coordinate at distance along Arc.
         """
+        if (distance is None) and (fraction is None) and (percentage is None):
+            raise Exception("You must provide either a distance, fraction or percentage.")
+
+        if (distance is not None) and (fraction is not None):
+            warn("Both distance and fraction provided. Using distance.", UserWarning)
+        if (distance is not None) and (percentage is not None):
+            warn("Both distance and percentage provided. Using distance.", UserWarning)
+
+        if distance is None:
+            if (fraction is not None) and (percentage is not None):
+                warn("Both fraction and percentage provided. Using fraction.", UserWarning)
+            if fraction is not None:
+                distance = self.length * fraction
+            elif percentage is not None:
+                distance = self.length * (percentage / 100)
+
         ref_coordinate_angle = atan2(
             (ref_coordinate.y - self.centre.y), (ref_coordinate.x - self.centre.x)
         )
@@ -1359,34 +2269,6 @@ class _BaseArc(Entity):
             distance / self.radius
         )  # sign of the radius accounts for clockwise/anticlockwise arcs
         return self.centre + Coordinate(*rt_to_xy(abs(self.radius), degrees(angle)))
-
-    def get_intersects_with_line(self, line, tolerance=0.0001):
-        """Get if a line intersects with this arc.
-
-        Parameters
-        ----------
-        line : ansys.motorcad.core.geometry.Line
-            Check if this Line entity intersects with self
-
-        Returns
-        -------
-        bool
-        """
-        start = self.start
-        end = self.end
-        points = 1 / tolerance  # 10000
-        length = 2 * pi * self.radius * (abs(self.total_angle) / 360)
-        line_crossed = False
-        for i in range(int(points)):
-            point_1 = self.get_coordinate_from_distance(start, (i * (length / points)))
-            point_2 = self.get_coordinate_from_distance(point_1, length / points)
-            tangent = Line(point_1, point_2)
-            intersection = tangent.get_line_intersection(line)
-            if intersection == None:
-                pass
-            elif tangent.coordinate_on_entity(intersection):
-                line_crossed = True
-        return line_crossed
 
     def mirror(self, mirror_line):
         """Mirror arc about a line.
@@ -1464,6 +2346,119 @@ class _BaseArc(Entity):
         return self.coordinate_within_arc_radius(coordinate) and isclose(
             abs(radius), abs(self.radius), abs_tol=GEOM_TOLERANCE
         )
+
+    def get_line_intersection(self, line):
+        """Get intersection Coordinates of arc with a line.
+
+        Returns None if intersection not found.
+
+        Parameters
+        ----------
+        line : ansys.motorcad.core.geometry.Line
+
+        Returns
+        -------
+        ansys.motorcad.core.geometry.Coordinate or list of Coordinate or None
+        """
+        # circle of the arc
+        a = self.centre.x
+        b = self.centre.y
+        r = self.radius
+
+        if line.is_vertical:
+            # line x coordinate is constant
+            x = line.start.x
+
+            A = 1
+            B = -2 * b
+            C = x**2 - 2 * x * a + a**2 + b**2 - r**2
+            D = B**2 - 4 * A * C
+
+            if D < 0:
+                # line and circle do not intersect
+                return None
+            elif D == 0:
+                # line and circle intersect at 1 point
+                y = -B / (2 * A)
+                return Coordinate(x, y)
+            else:
+                # line and circle intersect at 2 points
+                y1 = (-B + sqrt(D)) / (2 * A)
+                y2 = (-B - sqrt(D)) / (2 * A)
+                return [Coordinate(x, y1), Coordinate(x, y2)]
+        else:
+            # Normal case, line y is a function of x
+            m = line.gradient
+            c = line.y_intercept
+
+            A = 1 + m**2
+            B = 2 * (m * (c - b) - a)
+            C = a**2 + (c - b) ** 2 - r**2
+
+            D = B**2 - 4 * A * C
+            if D < 0:
+                # line and circle do not intersect
+                return None
+            elif D == 0:
+                # line and circle intersect at 1 point
+                x = -B / (2 * A)
+                y = m * x + c
+                return Coordinate(x, y)
+            else:
+                # line and circle intersect at 2 points
+                x1 = (-B + sqrt(D)) / (2 * A)
+                x2 = (-B - sqrt(D)) / (2 * A)
+                y1 = m * x1 + c
+                y2 = m * x2 + c
+                return [Coordinate(x1, y1), Coordinate(x2, y2)]
+
+    def get_arc_intersection(self, arc):
+        """Get intersection Coordinates of arc with another arc.
+
+        Returns None if intersection not found.
+
+        Parameters
+        ----------
+        arc : ansys.motorcad.core.geometry.Arc
+
+        Returns
+        -------
+        list of Coordinate or None
+        """
+        # circle of self
+        a1 = self.centre.x
+        b1 = self.centre.y
+        r1 = abs(self.radius)
+
+        # circle of other arc
+        a2 = arc.centre.x
+        b2 = arc.centre.y
+        r2 = abs(arc.radius)
+
+        d = sqrt((a2 - a1) ** 2 + (b2 - b1) ** 2)
+
+        if d > (r1 + r2):
+            # if they don't intersect
+            return None
+        if d < abs(r1 - r2):
+            # if one circle is inside the other
+            return None
+        if d == 0 and r1 == r2:
+            # coincident circles
+            return None
+        else:
+            a = (r1**2 - r2**2 + d**2) / (2 * d)
+            h = sqrt(r1**2 - a**2)
+
+            x = a1 + a * (a2 - a1) / d
+            y = b1 + a * (b2 - b1) / d
+            x1 = x + h * (b2 - b1) / d
+            y1 = y - h * (a2 - a1) / d
+
+            x2 = x - h * (b2 - b1) / d
+            y2 = y + h * (a2 - a1) / d
+
+            return [Coordinate(x1, y1), Coordinate(x2, y2)]
 
     @property
     def start_angle(self):
@@ -1648,6 +2643,22 @@ class Arc(_BaseArc):
             # Arc had an angle defined as greater than 180 deg
             return p_centre - Coordinate(l_x, l_y)
 
+    def get_bounds(self):
+        """
+        Return relevant information about the maximum and minimum x, y, and radius.
+
+        Returns: (max_radius, max_x, min_x, max_y, min_y)
+        """
+        max_radius = max(
+            Coordinate.get_polar_coords_deg(self.start)[0],
+            Coordinate.get_polar_coords_deg(self.end)[0],
+            Coordinate.get_polar_coords_deg(self.midpoint)[0],
+        )
+        xs = (self.start.x, self.end.x, self.midpoint.x)
+        ys = (self.start.y, self.end.y, self.midpoint.y)
+
+        return (max_radius, max(xs), min(xs), max(ys), min(ys))
+
 
 class EntityList(list):
     """Generic class for list of Entities."""
@@ -1656,6 +2667,10 @@ class EntityList(list):
         """Compare equality of 2 EntityList objects."""
         return self._entities_same(other, check_reverse=True)
 
+    def __ne__(self, other):
+        """Compare difference of 2 EntityList objects."""
+        return not self == other
+
     def reverse(self):
         """Reverse EntityList, including entity start end coordinates."""
         super().reverse()
@@ -1663,6 +2678,159 @@ class EntityList(list):
         # Also reverse entity start/end points so the EntityList is continuous
         for entity in self:
             entity.reverse()
+
+    @classmethod
+    def polygon(cls, points, sort=False):
+        """Create an EntityList from a list of points, connecting them with lines.
+
+        If sort is true, a centre point roughly in the centre of the supplied points will be
+        established and lines connecting the supplied points will be drawn anticlockwise about
+        that centre.
+
+        Parameters
+        ----------
+        points : list of Coordinates or tuples
+        sort : bool
+            Reorders points to make geometry valid:
+        """
+        # conversion to Coordinates:
+        if all(isinstance(point, tuple) for point in points):
+            points_new = [0] * len(points)
+            for i, p in enumerate(points):
+                points_new[i] = Coordinate(p[0], p[1])
+            points = points_new
+
+        if sort:
+            xcentre = sum(list(point.x for point in points)) / len(points)
+            ycentre = sum(list(point.y for point in points)) / len(points)
+
+            centre = Coordinate(xcentre, ycentre)
+
+            relative_points = list(deepcopy(point) for point in points)
+            for point in relative_points:
+                point.translate(-xcentre, -ycentre)
+
+            # To make sure behaviour is well-defined for points with an equal angular coordinate
+            # the points are first by sorted by radius, greatest to least
+            sorted_points_relative = sorted(
+                relative_points, key=lambda x: x.get_polar_coords_deg()[0], reverse=True
+            )
+
+            # Sort points by radial coordinate (relative to overall average position),
+            # from least to greatest angle (angles range from -180 to 180 degrees)
+            sorted_points_relative.sort(key=lambda x: x.get_polar_coords_deg()[1])
+
+            for point in sorted_points_relative:
+                point.translate(xcentre, ycentre)
+            points = sorted_points_relative
+
+        final_list = EntityList()
+
+        # Adds the lines
+        for count, point in enumerate(points):
+            if count == len(points) - 1:
+                final_list.append(Line(point, points[0]))
+            else:
+                final_list.append(Line(point, points[count + 1]))
+        if not final_list.has_valid_geometry:
+            warnings.warn("Entered point order may result in invalid geometry.")
+        return final_list
+
+    @property
+    def is_closed(self):
+        """Check whether entities create a closed region.
+
+        Returns
+        -------
+        Boolean
+            Whether EntityList is closed
+        """
+        if len(self) > 0:
+            # Make sure all adjacent entities have a point in common
+            return all(
+                get_entities_have_common_coordinate(self[i - 1], self[i])
+                for i in range(0, len(self))
+            )
+        else:
+            return False
+
+    @property
+    def self_intersecting(self):
+        """Check whether entities intersect each other.
+
+        Returns
+        -------
+        Boolean
+            Whether entities are intersecting
+        """
+        entity_pairs = []
+        # Create a set of all non-neighbor pairs of lines/arcs in the region
+        for entity in self:
+            for entity_other in self:
+                if entity != entity_other:
+                    if (
+                        entity.start != entity_other.end and entity.end != entity_other.start
+                    ) and not (
+                        (entity, entity_other) in entity_pairs
+                        or (entity_other, entity) in entity_pairs
+                    ):
+                        entity_pairs.append((entity, entity_other))
+
+        # Check if they have any intersections
+        for entity1, entity2 in entity_pairs:
+            intersections = entity1.get_intersection(entity2)
+            if intersections is not None:
+                for intersection in intersections:
+                    if entity1.coordinate_on_entity(intersection) and entity2.coordinate_on_entity(
+                        intersection
+                    ):
+                        return True
+        return False
+
+    @property
+    def is_anticlockwise(self):
+        """Check whether EntityList is connected in an anticlockwise manner.
+
+        Returns
+        -------
+        Boolean
+            Whether EntityList is anticlockwise
+        """
+        # Checks to make sure checking direction even makes sense
+        if not (not self.self_intersecting and self.is_closed):
+            raise Exception("Entities must be closed and nonintersecting")
+
+        # Find the lowest point, as well the entities coming in and out of that point
+        points = sorted(self.points, key=lambda p: p.y)
+        lowest = points[0]
+
+        for entity in self:
+            if entity.start == lowest:
+                ent_out = entity
+            if entity.end == lowest:
+                ent_in = entity
+
+        # Determine the  angles of the entry and exit lines (arcs can make this tricky,
+        # so the angle is determined from the start and a point very near it)
+        exit_angle = Line(
+            ent_out.start, ent_out.get_coordinate_from_distance(ent_out.start, fraction=0.001)
+        ).angle
+        entry_angle = Line(
+            ent_in.end, ent_in.get_coordinate_from_distance(ent_in.end, fraction=0.001)
+        ).angle
+
+        return entry_angle > exit_angle
+
+    @property
+    def has_valid_geometry(self):
+        """Check whether geometry is valid for motorcad.
+
+        Returns
+        -------
+        Boolean
+            Whether geometry is valid for motorcad
+        """
+        return self.is_closed and (not self.self_intersecting) and self.is_anticlockwise
 
     @property
     def points(self):
@@ -1711,6 +2879,9 @@ class EntityList(list):
                 return True
             else:
                 return False
+
+        if len(self) != len(entities_to_compare):
+            return False
 
         if check_reverse:
             if _entities_same_with_direction(self, entities_to_compare):
@@ -1893,6 +3064,7 @@ def _orientation_of_three_points(c1, c2, c3):
         Coordinate 2
     c3 : ansys.motorcad.core.geometry.Coordinate
         Coordinate 3
+
     Returns
     -------
         _Orientation
@@ -1910,61 +3082,59 @@ def _orientation_of_three_points(c1, c2, c3):
         return _Orientation.collinear
 
 
-class RegionType(Enum):
-    """Provides an enumeration for storing Motor-CAD region types."""
+def _bernstein(n, v, t):
+    """Find the Bernstein polynomial value.
 
-    stator = "Stator"
-    rotor = "Rotor"
-    slot_area_stator = "Stator Slot"
-    slot_area_rotor = "Rotor Slot"
-    slot_split = "Split Slot"
-    stator_liner = "Stator Liner"
-    rotor_liner = "Rotor Liner"
-    wedge = "Wedge"
-    stator_duct = "Stator Duct"
-    housing = "Housing"
-    housing_magnetic = "Magnetic Housing"
-    stator_impreg = "Stator Impreg"
-    impreg_gap = "Impreg Gap"
-    stator_copper = "Stator Copper"
-    stator_copper_ins = "Stator Copper Insulation"
-    stator_divider = "Stator Divider"
-    stator_slot_spacer = "Stator Slot Spacer"
-    stator_separator = "Stator slot separator"
-    coil_insulation = "Coil Insulation"
-    stator_air = "Stator Air"
-    rotor_hub = "Rotor hub"
-    rotor_air = "Rotor Air"
-    rotor_air_exc_liner = "Rotor Air (excluding liner area)"
-    rotor_pocket = "Rotor Pocket"
-    pole_spacer = "Pole Spacer"
-    rotor_slot = "Rotor Slot"
-    coil_separator = "Coil Separator"
-    damper_bar = "Damper Bar"
-    wedge_rotor = "Rotor Wedge"
-    rotor_divider = "Rotor Divider"
-    rotor_copper_ins = "Rotor Copper Insulation"
-    rotor_copper = "Rotor Copper"
-    rotor_impreg = "Rotor Impreg"
-    shaft = "Shaft"
-    axle = "Axle"
-    rotor_duct = "Rotor Duct"
-    magnet = "Magnet"
-    barrier = "Barrier"
-    mounting_base = "Base Mount"
-    mounting_plate = "Plate Mount"
-    banding = "Banding"
-    sleeve = "Sleeve"
-    rotor_cover = "Rotor Cover"
-    slot_wj_insulation = "Slot Water Jacket Insulation"
-    slot_wj_wall = "Slot Water Jacket Wall"
-    slot_wj_duct = "Slot Water Jacket Duct"
-    slot_wj_duct_no_detail = "Slot Water Jacket Duct (no detail)"
-    cowling = "Cowling"
-    cowling_gril = "Cowling Grill"
-    brush = "Brush"
-    commutator = "Commutator"
-    airgap = "Airgap"
-    dxf_import = "DXF Import"
-    impreg_loss_lot_ac_loss = "Stator Proximity Loss Slot"
-    adaptive = "Adaptive Region"
+    Parameters
+    ----------
+    n : int
+        Degree of Bernstein polynomial
+    v : int
+        Index of Bernstein polynomial
+    t : float
+        Value to calculate in range 0-1
+    Returns
+    -------
+        float
+    """
+    return comb(n, v) * t**v * (1 - t) ** (n - v)
+
+
+def get_bezier_points(control_points, num_output_points):
+    """Find a list of coordinates along a Bezier curve.
+
+    Parameters
+    ----------
+    control points : List of Coordinate
+        The control points for the Bezier curve
+    num_output_points : int
+        The number of samples along the curve
+
+    Returns
+    -------
+        List of Coordinate
+    """
+    points = []
+    num_control = len(control_points)
+
+    for output_point in range(num_output_points):
+        t = float(output_point) / (num_output_points - 1)
+        bernstein_values = []
+
+        for control_point_index in range(num_control):
+            bernstein_values.append(_bernstein(num_control - 1, control_point_index, t))
+
+        point = []
+        # Coordinates are two-dimensional
+        for dimension in range(2):
+            value = 0
+            for control_point_index, control_point in enumerate(control_points):
+                if dimension == 0:
+                    control_point_value = control_point.x
+                else:
+                    control_point_value = control_point.y
+                value = value + control_point_value * bernstein_values[control_point_index]
+            point.append(value)
+        coordinate = Coordinate(point[0], point[1])
+        points.append(coordinate)
+    return points
