@@ -1,4 +1,4 @@
-# Copyright (C) 2022 - 2024 ANSYS, Inc. and/or its affiliates.
+# Copyright (C) 2022 - 2026 ANSYS, Inc. and/or its affiliates.
 # SPDX-License-Identifier: MIT
 #
 #
@@ -24,7 +24,8 @@
 from warnings import warn
 
 from ansys.motorcad.core.geometry import Region, RegionMagnet
-from ansys.motorcad.core.rpc_client_core import is_running_in_internal_scripting
+from ansys.motorcad.core.geometry_tree import GeometryTree
+from ansys.motorcad.core.rpc_client_core import MotorCADError, is_running_in_internal_scripting
 
 
 class _RpcMethodsAdaptiveGeometry:
@@ -64,13 +65,31 @@ class _RpcMethodsAdaptiveGeometry:
         params = [name]
         return self.connection.send_and_receive(method, params)
 
-    def get_region(self, name):
+    def set_adaptive_parameter_default(self, name, value):
+        """Set default value for an adaptive parameter, if the parameter does not already exist.
+
+        Parameters
+        ----------
+        name : string
+            name of parameter.
+        value : float
+            value of parameter.
+        """
+        self.connection.ensure_version_at_least("2024.0")
+        try:
+            self.get_adaptive_parameter_value(name)
+        except MotorCADError:
+            self.set_adaptive_parameter_value(name, value)
+
+    def get_region(self, name, get_linked=False):
         """Get Motor-CAD geometry region.
 
         Parameters
         ----------
         name : string
             name of region.
+        get_linked : boolean
+            if linked regions should also be returned
 
         Returns
         -------
@@ -84,7 +103,10 @@ class _RpcMethodsAdaptiveGeometry:
         raw_region = self.connection.send_and_receive(method, params)
 
         region = Region._from_json(raw_region, motorcad_instance=self)
-
+        if get_linked:
+            self.connection.ensure_version_at_least("2026.0")
+            for linked_region in raw_region["linked_regions"]:
+                region.linked_regions.append(self.get_region(linked_region, get_linked=False))
         return region
 
     def get_region_dxf(self, name):
@@ -288,3 +310,29 @@ class _RpcMethodsAdaptiveGeometry:
         # No need to do this if running internally
         if not is_running_in_internal_scripting():
             return self.connection.send_and_receive(method)
+
+    def get_geometry_tree(self):
+        """Fetch a GeometryTree object containing all the defining geometry of the loaded motor.
+
+        Returns
+        -------
+        ansys.motorcad.core.geometry_tree.GeometryTree
+            Motor-CAD geometry tree
+        """
+        self.connection.ensure_version_at_least("2026.0")
+        method = "GetGeometryTree"
+        json = self.connection.send_and_receive(method)
+        return GeometryTree._from_json(json, self)
+
+    def set_geometry_tree(self, tree: GeometryTree):
+        """Use a GeometryTree object to set the defining geometry of the loaded motor."""
+        self.connection.ensure_version_at_least("2026.0")
+        params = [tree._to_json()]
+        method = "SetGeometryTree"
+        return self.connection.send_and_receive(method, params)
+
+    def get_maxwell_udm_geometry_json(self):
+        """Fetch a dict defining Maxwell UDM geometry."""
+        self.connection.ensure_version_at_least("2026.0")
+        method = "GetGeometryTree_Maxwell_UDM"
+        return self.connection.send_and_receive(method)
