@@ -1,4 +1,4 @@
-# Copyright (C) 2022 - 2024 ANSYS, Inc. and/or its affiliates.
+# Copyright (C) 2022 - 2026 ANSYS, Inc. and/or its affiliates.
 # SPDX-License-Identifier: MIT
 #
 #
@@ -21,6 +21,7 @@
 # SOFTWARE.
 
 """RPC methods (general)."""
+from ansys.motorcad.core.rpc_client_core import MotorCADError
 
 
 class _RpcMethodsGeneral:
@@ -378,31 +379,49 @@ class _RpcMethodsGeneral:
         return self.connection.send_and_receive(method, params)
 
     def save_results(self, solution_type):
-        """Save the output results from an ``"EMagnetic"`` solution.
+        """Save the output results from an ``"EMagnetic"`` or ``"Thermal"`` solution.
 
-        This method supports only ``"EMagnetic"`` solutions.
+        This method supports only ``"EMagnetic"`` or ``"Thermal"`` solutions.
 
         Parameters
         ----------
         solution_type : str
-            Soultion type, which must be ``"EMagnetic"``.
+            Solution type, which must be ``"EMagnetic"`` or ``"Thermal"``.
         """
         method = "SaveResults"
         params = [solution_type]
+
+        if solution_type.lower() == "thermal":
+            self.connection.ensure_version_at_least("2025.0")
+        elif solution_type.lower() != "emagnetic":
+            raise MotorCADError(
+                "Save results are not available for this solution type: "
+                + solution_type
+                + "\nAvailable solution types are: Thermal, EMagnetic"
+            )
         return self.connection.send_and_receive(method, params)
 
     def load_results(self, solution_type):
-        """Load the output results from an ``"EMagnetic"`` solution.
+        """Load the output results from an ``"EMagnetic"`` or ``"Thermal"`` solution.
 
-        This method supports only ``"EMagnetic"`` solution.
+        This method supports only ``"EMagnetic"`` or ``"Thermal"`` solution.
 
         Parameters
         ----------
         solution_type : str
-            Soultion type, which must be ``"EMagnetic"``.
+            Solution type, which must be ``"EMagnetic"`` or ``"Thermal"``.
         """
         method = "LoadResults"
         params = [solution_type]
+
+        if solution_type.lower() == "thermal":
+            self.connection.ensure_version_at_least("2025.0")
+        elif solution_type.lower() != "emagnetic":
+            raise MotorCADError(
+                "Load results are not available for this solution type: "
+                + solution_type
+                + "\nAvailable solution types are: Thermal, EMagnetic"
+            )
         return self.connection.send_and_receive(method, params)
 
     def load_magnetisation_curves(self, file_path):
@@ -525,3 +544,29 @@ class _RpcMethodsGeneral:
         method = "SetCurrentFileJSON"
         params = [file_contents]
         self.connection.send_and_receive(method, params)
+
+    def load_nvh_custom_response(self, file_name):
+        """Load custom characteristic noise response functions from space/tab delimited text file.
+
+        Parameters
+        ----------
+        file_name: str
+            Full path to the text file
+        """
+        self.connection.ensure_version_at_least("2026.0")
+        method = "LoadNVHResponse"
+        params = [file_name]
+        return self.connection.send_and_receive(method, params)
+
+    def save_nvh_custom_response(self, file_name):
+        """Save custom characteristic noise response functions to tab delimited text file.
+
+        Parameters
+        ----------
+        file_name: str
+            Full path for the text file
+        """
+        self.connection.ensure_version_at_least("2026.0")
+        method = "SaveNVHResponse"
+        params = [file_name]
+        return self.connection.send_and_receive(method, params)
