@@ -1,4 +1,4 @@
-# Copyright (C) 2022 - 2024 ANSYS, Inc. and/or its affiliates.
+# Copyright (C) 2022 - 2026 ANSYS, Inc. and/or its affiliates.
 # SPDX-License-Identifier: MIT
 #
 #
@@ -22,6 +22,8 @@
 
 """RPC methods for variables."""
 from warnings import warn
+
+from ansys.motorcad.core.datastore import Datastore
 
 
 class _RpcMethodsVariables:
@@ -146,9 +148,33 @@ class _RpcMethodsVariables:
         str
             Current .mot file path and name
         """
-        method = "GetMotorCADFileName"
-        if self.connection.send_and_receive(method) == "":
-            warn("No file has been loaded in this MotorCAD instance")
-            return None
+        if self.connection.check_version_at_least("2025.0"):
+            method = "GetMotorCADFileName"
+            if self.connection.send_and_receive(method) == "":
+                warn("No file has been loaded in this MotorCAD instance")
+                return None
+            else:
+                return self.connection.send_and_receive(method)
         else:
-            return self.connection.send_and_receive(method)
+            warn(
+                "GetMotorCADFileName not available in Motor-CAD "
+                + self.connection.program_version
+                + ". Returning value of CurrentMotFilePath_MotorLAB"
+            )
+            if self.get_variable("CurrentMotFilePath_MotorLAB") == "":
+                warn("No file has been loaded in this MotorCAD instance")
+                return None
+            else:
+                return self.get_variable("CurrentMotFilePath_MotorLAB")
+
+    def get_datastore(self):
+        """Get the whole database from Motor-CAD.
+
+        Returns
+        -------
+        ansys.motorcad.core.datastore.Datastore
+        """
+        self.connection.ensure_version_at_least("2026.0")
+        method = "GetDataStore"
+        datastore_json = self.connection.send_and_receive(method)
+        return Datastore.from_json(datastore_json)
