@@ -1,4 +1,4 @@
-# Copyright (C) 2022 - 2024 ANSYS, Inc. and/or its affiliates.
+# Copyright (C) 2022 - 2026 ANSYS, Inc. and/or its affiliates.
 # SPDX-License-Identifier: MIT
 #
 #
@@ -19,15 +19,11 @@
 # LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
-
 import pytest
 
 from RPC_Test_Common import reset_temp_file_folder, reset_to_default_file
 import ansys.motorcad.core
 from ansys.motorcad.core import MotorCAD
-
-motorcad_instance = None
-motorcad_instance_fea_old = None
 
 
 def pytest_sessionstart(session):
@@ -35,40 +31,34 @@ def pytest_sessionstart(session):
     ansys.motorcad.core.rpc_client_core.DONT_CHECK_MOTORCAD_VERSION = True
 
 
-def pytest_sessionfinish(session, exitstatus):
-    """
-    Called after whole test run finished, right before
-    returning the exit status to the system.
-    """
-    ...  # Your code goes here
-    global motorcad_instance
-    if hasattr(motorcad_instance, "quit"):
-        motorcad_instance.quit()
-
-
-@pytest.fixture
+@pytest.fixture(scope="session")
 def mc():
     """Set up test environment for whole unit of tests"""
-    global motorcad_instance
+    motorcad_instance = MotorCAD()
+    # Disable messages if opened with UI
+    motorcad_instance.set_variable("MessageDisplayState", 2)
+    reset_to_default_file(motorcad_instance)
 
-    if not (hasattr(motorcad_instance, "is_open") and motorcad_instance.is_open()):
-        motorcad_instance = MotorCAD()
-        # Disable messages if opened with UI
-        motorcad_instance.set_variable("MessageDisplayState", 2)
-        reset_to_default_file(motorcad_instance)
+    yield motorcad_instance
 
-    return motorcad_instance
+    motorcad_instance.quit()
 
 
-@pytest.fixture
+@pytest.fixture(scope="function")
+def mc_reset_to_default_on_teardown(mc):
+    """Set up test environment for whole unit of tests"""
+    yield mc
+    reset_to_default_file(mc)
+
+
+@pytest.fixture(scope="session")
 def mc_fea_old():
     """Old fea geometry tests cause lots of conflicts - use a new MotorCAD"""
-    global motorcad_instance_fea_old
+    motorcad_instance_fea_old = MotorCAD()
+    # Disable messages if opened with UI
+    motorcad_instance_fea_old.set_variable("MessageDisplayState", 2)
+    reset_to_default_file(motorcad_instance_fea_old)
 
-    if not (hasattr(motorcad_instance_fea_old, "is_open") and motorcad_instance_fea_old.is_open()):
-        motorcad_instance_fea_old = MotorCAD()
-        # Disable messages if opened with UI
-        motorcad_instance_fea_old.set_variable("MessageDisplayState", 2)
-        reset_to_default_file(motorcad_instance_fea_old)
+    yield motorcad_instance_fea_old
 
-    return motorcad_instance_fea_old
+    motorcad_instance_fea_old.quit()
