@@ -1,4 +1,4 @@
-# Copyright (C) 2022 - 2024 ANSYS, Inc. and/or its affiliates.
+# Copyright (C) 2022 - 2026 ANSYS, Inc. and/or its affiliates.
 # SPDX-License-Identifier: MIT
 #
 #
@@ -22,13 +22,33 @@
 
 import os
 
-from RPC_Test_Common import get_temp_files_dir_path
+import pytest
+
+from RPC_Test_Common import get_temp_files_dir_path, reset_to_default_file
+from ansys.motorcad.core import MotorCAD
 
 
-def test_show_message(mc):
+@pytest.fixture(scope="module")
+def mc_ui():
+    """Set up test environment for whole unit of tests"""
+
+    os.environ["MOTORCAD_FORCE_SHOW_GUI"] = "1"
+    try:
+        motorcad_instance_ui = MotorCAD()
+    finally:
+        os.environ.pop("MOTORCAD_FORCE_SHOW_GUI")
+
+    # Disable messages if opened with UI
+    motorcad_instance_ui.set_variable("MessageDisplayState", 2)
+    reset_to_default_file(motorcad_instance_ui)
+
+    return motorcad_instance_ui
+
+
+def test_show_message(mc_ui):
     test_message = "test 1"
-    mc.show_message(test_message)
-    messages = mc.get_messages(1)
+    mc_ui.show_message(test_message)
+    messages = mc_ui.get_messages(1)
     assert (test_message in message for message in messages)
 
 
@@ -36,50 +56,59 @@ def test_show_message(mc):
 # These might raise an exception if they fail
 
 
-def test_show_magnetic_context(mc):
-    mc.show_magnetic_context()
+def test_show_magnetic_context(mc_ui):
+    mc_ui.show_magnetic_context()
 
 
-def test_show_mechanical_context(mc):
-    mc.show_mechanical_context()
+def test_show_mechanical_context(mc_ui):
+    mc_ui.show_mechanical_context()
 
 
-def test_show_thermal_context(mc):
-    mc.show_thermal_context()
+def test_show_thermal_context(mc_ui):
+    mc_ui.show_thermal_context()
 
 
-def test_set_motorlab_context(mc):
-    mc.set_motorlab_context()
+def test_set_motorlab_context(mc_ui):
+    mc_ui.set_motorlab_context()
 
 
-def test_save_screen_to_file(mc):
+def test_save_screen_to_file(mc_ui):
     file_path = get_temp_files_dir_path() + r"\screen.png"
 
-    mc.save_screen_to_file("Radial", file_path)
+    mc_ui.save_screen_to_file("Radial", file_path)
 
     assert os.path.exists(file_path)
 
 
-def test_display_screen(mc):
+def test_display_screen(mc_ui):
     # Difficult to check this has actually worked
     # These might raise an exception if they fail
-    mc.initialise_tab_names()
-    mc.display_screen("Calculation")
+    mc_ui.initialise_tab_names()
+    mc_ui.display_screen("Calculation")
 
 
-def test_set_3d_component_visibility(mc):
-    mc.set_3d_component_visibility("stator", "winding", 0)
+def test_set_3d_component_visibility(mc_ui):
+    mc_ui.set_3d_component_visibility("stator", "winding", 0)
 
 
-def test_set_visible(mc):
-    mc.set_visible(False)
-    mc.set_visible(True)
+def test_set_visible(mc_ui):
+    mc_ui.set_visible(False)
+    mc_ui.set_visible(True)
 
 
-def test_clear_messages(mc):
+def test_clear_messages(mc_ui):
     test_message = "test_message_to_clear"
-    mc.show_message(test_message)
-    assert (test_message in message for message in mc.get_messages(0))
+    mc_ui.show_message(test_message)
+    assert (test_message in message for message in mc_ui.get_messages(0))
 
-    mc.clear_messages()
-    assert mc.get_messages(0) == [""]
+    mc_ui.clear_messages()
+
+    messages = mc_ui.get_messages(0)
+
+    if messages == [""]:
+        assert True
+    elif (len(messages) == 1) and ("pymotorcad: get_messages" in messages[0]):
+        # get_messages might get logged if using Motor-CAD in debug mode
+        assert True
+    else:
+        assert False
