@@ -1,4 +1,4 @@
-# Copyright (C) 2022 - 2024 ANSYS, Inc. and/or its affiliates.
+# Copyright (C) 2022 - 2026 ANSYS, Inc. and/or its affiliates.
 # SPDX-License-Identifier: MIT
 #
 #
@@ -22,6 +22,8 @@
 
 import os
 
+import pytest
+
 from RPC_Test_Common import (
     almost_equal,
     get_base_test_file_path,
@@ -31,7 +33,7 @@ from RPC_Test_Common import (
     reset_to_default_file,
 )
 import ansys.motorcad.core
-from ansys.motorcad.core import MotorCAD
+from ansys.motorcad.core import MotorCAD, MotorCADError
 
 # Allows us to add a new api method to testing before the next Motor-CAD release is available
 # Dev release will have a lower version number than actual release so don't want to check this
@@ -42,30 +44,29 @@ def kh_to_ms(kh):
     return kh * 0.2777778
 
 
-def test_save_load_clear_duty_cycle(mc):
+def test_save_load_clear_duty_cycle(mc_reset_to_default_on_teardown):
     # thermal transient calc
-    mc.set_variable("ThermalCalcType", 1)
+    mc_reset_to_default_on_teardown.set_variable("ThermalCalcType", 1)
 
     # transient calc type - duty cycle
-    mc.set_variable("TransientCalculationType", 1)
+    mc_reset_to_default_on_teardown.set_variable("TransientCalculationType", 1)
 
     assert almost_equal(
-        mc.get_array_variable("Duty_Cycle_Vehicle_Speed_Start", 20), kh_to_ms(64.21)
+        mc_reset_to_default_on_teardown.get_array_variable("Duty_Cycle_Vehicle_Speed_Start", 20),
+        kh_to_ms(64.21),
     )
 
     save_path = get_temp_files_dir_path() + r"\duty_cycle.dat"
-    mc.save_duty_cycle(save_path)
+    mc_reset_to_default_on_teardown.save_duty_cycle(save_path)
 
-    mc.clear_duty_cycle()
-    assert mc.get_variable("Duty_Cycle_Num_Periods") < 2
+    mc_reset_to_default_on_teardown.clear_duty_cycle()
+    assert mc_reset_to_default_on_teardown.get_variable("Duty_Cycle_Num_Periods") < 2
 
-    mc.load_duty_cycle(save_path)
+    mc_reset_to_default_on_teardown.load_duty_cycle(save_path)
     assert almost_equal(
-        mc.get_array_variable("Duty_Cycle_Vehicle_Speed_Start", 20), kh_to_ms(64.21)
+        mc_reset_to_default_on_teardown.get_array_variable("Duty_Cycle_Vehicle_Speed_Start", 20),
+        kh_to_ms(64.21),
     )
-
-    # reset model state
-    reset_to_default_file(mc)
 
 
 def test_export_matrices(mc):
@@ -143,13 +144,11 @@ def test_export_force_animation(mc):
     mc.export_force_animation("Radial", file_path)
 
 
-def test_load_template(mc):
-    mc.load_template("e5")
+def test_load_template(mc_reset_to_default_on_teardown):
+    mc_reset_to_default_on_teardown.load_template("e5")
 
     # Check we have loaded IM motor
-    assert mc.get_variable("Motor_Type") == 1
-
-    reset_to_default_file(mc)
+    assert mc_reset_to_default_on_teardown.get_variable("Motor_Type") == 1
 
 
 def test_export_multi_force_data(mc):
@@ -169,40 +168,67 @@ def test_geometry_export(mc):
     assert os.path.exists(file_path)
 
 
-def test_save_load_magnetisation_curves(mc):
-    mc.load_from_file(get_dir_path() + r"\test_files\SRM_test_file.mot")
-    mc.do_magnetic_calculation()
+def test_save_load_magnetisation_curves(mc_reset_to_default_on_teardown):
+    mc_reset_to_default_on_teardown.load_from_file(
+        get_dir_path() + r"\test_files\SRM_test_file.mot"
+    )
+    mc_reset_to_default_on_teardown.do_magnetic_calculation()
 
     file_path = get_temp_files_dir_path() + r"\mag_curves.txt"
 
-    mc.save_magnetisation_curves(file_path)
+    mc_reset_to_default_on_teardown.save_magnetisation_curves(file_path)
     assert os.path.exists(file_path)
 
-    mc.set_variable("MaxCurrent_MagnetisationCurves", 20)
-    mc.do_magnetic_calculation()
-    assert mc.get_variable("MaxCurrent_Calculated_MagnetisationCurves") == 20
+    mc_reset_to_default_on_teardown.set_variable("MaxCurrent_MagnetisationCurves", 20)
+    mc_reset_to_default_on_teardown.do_magnetic_calculation()
+    assert (
+        mc_reset_to_default_on_teardown.get_variable("MaxCurrent_Calculated_MagnetisationCurves")
+        == 20
+    )
 
-    mc.load_magnetisation_curves(file_path)
-    assert mc.get_variable("MaxCurrent_Calculated_MagnetisationCurves") == 5
+    mc_reset_to_default_on_teardown.load_magnetisation_curves(file_path)
+    assert (
+        mc_reset_to_default_on_teardown.get_variable("MaxCurrent_Calculated_MagnetisationCurves")
+        == 5
+    )
 
-    reset_to_default_file(mc)
 
-
-def test_save_load_results():
+def test_save_load_results(mc_reset_to_default_on_teardown):
     # Currently not working as part of full tests
     # Works individually - need to look into this
-    pass
-    # mc.do_magnetic_calculation()
-    # mc.save_results("EMagnetic")
-    # assert os.path.exists(get_temp_files_dir_path() + r"\temp_test_file\EMag\outputResults.mot")
-    # assert os.path.exists(get_temp_files_dir_path() + r"\temp_test_file\EMag\GraphResults.ini")
-    #
-    # mc.load_from_file(get_temp_files_dir_path() + r"\temp_test_file.mot")
-    #
-    # mc.load_results("EMagnetic")
-    # assert mc.get_variable("MaxTorque") != 0
-    #
-    # reset_to_default_file(mc)
+
+    # EMag test
+    mc_reset_to_default_on_teardown.do_magnetic_calculation()
+    mc_reset_to_default_on_teardown.save_results("EMagnetic")
+    assert os.path.exists(get_temp_files_dir_path() + r"\temp_test_file\EMag\outputResults.mot")
+    assert os.path.exists(get_temp_files_dir_path() + r"\temp_test_file\EMag\GraphResults.ini")
+
+    mc_reset_to_default_on_teardown.load_from_file(
+        get_temp_files_dir_path() + r"\temp_test_file.mot"
+    )
+
+    mc_reset_to_default_on_teardown.load_results("EMagnetic")
+    assert mc_reset_to_default_on_teardown.get_variable("MaxTorque") != 0
+
+    reset_to_default_file(mc_reset_to_default_on_teardown)
+
+    # Thermal test - transient graphs only
+    mc_reset_to_default_on_teardown.do_transient_analysis()
+    mc_reset_to_default_on_teardown.save_results("Thermal")
+    assert os.path.exists(get_temp_files_dir_path() + r"\temp_test_file\Thermal\GraphResults.ini")
+
+    mc_reset_to_default_on_teardown.load_from_file(
+        get_temp_files_dir_path() + r"\temp_test_file.mot"
+    )
+
+    mc_reset_to_default_on_teardown.load_results("TheRmal")
+    assert mc_reset_to_default_on_teardown.get_power_graph_point("Armature Copper(Total)", 5) != 0
+
+    with pytest.raises(MotorCADError):
+        mc_reset_to_default_on_teardown.load_results("wrong_type")
+
+    with pytest.raises(MotorCADError):
+        mc_reset_to_default_on_teardown.save_results("wrong_type")
 
 
 def test_get_message(mc):
@@ -237,33 +263,76 @@ def file_line_differences(file_1, file_2):
     return number_differences
 
 
-def test_download_mot_file(mc):
+def test_download_mot_file(mc_reset_to_default_on_teardown):
     # Load and save base file so that contents are updated for this version of Motor-CAD
-    mc.load_from_file(get_base_test_file_path())
+    mc_reset_to_default_on_teardown.load_from_file(get_base_test_file_path())
     save_file_path = get_temp_files_dir_path() + r"\base_test_file_copy.mot"
-    mc.save_to_file(save_file_path)
+    mc_reset_to_default_on_teardown.save_to_file(save_file_path)
 
     download_mot_file_path = get_temp_files_dir_path() + r"\download_test_file.mot"
-    mc.download_mot_file(download_mot_file_path)
+    mc_reset_to_default_on_teardown.download_mot_file(download_mot_file_path)
 
     # File contents should be identical since downloading file without modifying
     assert file_contents_equal(save_file_path, download_mot_file_path)
-    reset_to_default_file(mc)
 
 
-def test_upload_mot_file(mc):
+def test_upload_mot_file(mc_reset_to_default_on_teardown):
     # Load and save base file so that contents are updated for this version of Motor-CAD
-    mc.load_from_file(get_test_files_dir_path() + r"\IM_test_file.mot")
+    mc_reset_to_default_on_teardown.load_from_file(get_test_files_dir_path() + r"\IM_test_file.mot")
     save_file_path = get_temp_files_dir_path() + r"\IM_test_file_copy.mot"
-    mc.save_to_file(save_file_path)
+    mc_reset_to_default_on_teardown.save_to_file(save_file_path)
 
     IM_test_file_path = get_test_files_dir_path() + r"\IM_test_file.mot"
-    mc.upload_mot_file(IM_test_file_path)
+    mc_reset_to_default_on_teardown.upload_mot_file(IM_test_file_path)
 
     upload_mot_file_path = get_temp_files_dir_path() + r"\upload_test_file.mot"
-    mc.save_to_file(upload_mot_file_path)
+    mc_reset_to_default_on_teardown.save_to_file(upload_mot_file_path)
 
     # File might have slight differences (paths etc.) since we are uploading and
     # saving as a new file
-    assert file_line_differences(upload_mot_file_path, save_file_path) < 30
-    reset_to_default_file(mc)
+    # 30 line differences is not enough if running on self-hosted runner.
+    # materials database sometimes saves copies of materials so we can get many more lines.
+    # files are ~14000 lines long so 200 is still a small amount of changes
+    assert file_line_differences(upload_mot_file_path, save_file_path) < 200
+
+
+def test_save_load_custom_response(mc_reset_to_default_on_teardown):
+    # set data
+    mc_reset_to_default_on_teardown.set_variable("CustomNVHFRFOrders", 2)
+    mc_reset_to_default_on_teardown.set_variable("CustomNVHFRFPoints", 3)
+
+    mc_reset_to_default_on_teardown.set_array_variable("CustomNVHFRFModeShape", 0, 0)
+    mc_reset_to_default_on_teardown.set_array_variable("CustomNVHFRFModeExcitation", 0, "R")
+
+    mc_reset_to_default_on_teardown.set_array_variable("CustomNVHFRFModeShape", 1, 0)
+    mc_reset_to_default_on_teardown.set_array_variable("CustomNVHFRFModeExcitation", 1, "T")
+
+    # Frequencies
+    mc_reset_to_default_on_teardown.set_array_variable("CustomNVHFRFFrequencies", 0, 20)
+    mc_reset_to_default_on_teardown.set_array_variable("CustomNVHFRFFrequencies", 1, 100)
+    mc_reset_to_default_on_teardown.set_array_variable("CustomNVHFRFFrequencies", 2, 500)
+
+    # FRF 0 Radial
+    mc_reset_to_default_on_teardown.set_array_variable_2d("CustomNVHFRFResponse", 0, 0, 0)
+    mc_reset_to_default_on_teardown.set_array_variable_2d("CustomNVHFRFResponse", 1, 0, 10)
+    mc_reset_to_default_on_teardown.set_array_variable_2d("CustomNVHFRFResponse", 2, 0, 20)
+
+    # FRF 0 Tangential
+    mc_reset_to_default_on_teardown.set_array_variable_2d("CustomNVHFRFResponse", 0, 1, 40)
+    mc_reset_to_default_on_teardown.set_array_variable_2d("CustomNVHFRFResponse", 1, 1, 60)
+    mc_reset_to_default_on_teardown.set_array_variable_2d("CustomNVHFRFResponse", 2, 1, 80)
+
+    file_path = get_dir_path() + r"\test_files\SaveLoadNVHResponse.txt"
+    if os.path.exists(file_path):
+        os.remove(file_path)
+
+    # save to file, clear and re-load
+    mc_reset_to_default_on_teardown.save_nvh_custom_response(file_path)
+    reset_to_default_file(mc_reset_to_default_on_teardown)
+    mc_reset_to_default_on_teardown.load_nvh_custom_response(file_path)
+
+    assert mc_reset_to_default_on_teardown.get_array_variable_2d("CustomNVHFRFResponse", 2, 1) == 80
+    assert mc_reset_to_default_on_teardown.get_array_variable("CustomNVHFRFModeShape", 1) == 0
+    assert (
+        mc_reset_to_default_on_teardown.get_array_variable("CustomNVHFRFModeExcitation", 1) == "T"
+    )
