@@ -1,4 +1,4 @@
-# Copyright (C) 2022 - 2024 ANSYS, Inc. and/or its affiliates.
+# Copyright (C) 2022 - 2026 ANSYS, Inc. and/or its affiliates.
 # SPDX-License-Identifier: MIT
 #
 #
@@ -19,14 +19,17 @@
 # LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
-from matplotlib import pyplot as plt
+import os
+import tempfile
+
+import matplotlib
+
+matplotlib.use("Agg")
 import pytest
 
-from ansys.motorcad.core.geometry import Coordinate, Line, Region
-import ansys.motorcad.core.geometry_drawing
-from ansys.motorcad.core.geometry_drawing import draw_objects
-
-# from ansys.motorcad.core.rpc_client_core import DEFAULT_INSTANCE, set_default_instance
+import ansys
+from ansys.motorcad.core.geometry import Coordinate, Line, Region, RegionType
+from ansys.motorcad.core.geometry_drawing import BiDict, draw_objects
 
 drawing_flag = False
 
@@ -37,7 +40,7 @@ def set_drawing_flag(*args, **kwargs):
 
 
 def create_triangle_reg(bottom_left_coord):
-    region = Region()
+    region = Region(RegionType.stator_air)
     region.name = "region " + str(bottom_left_coord)
     c1 = bottom_left_coord
     c2 = bottom_left_coord + Coordinate(10, 0)
@@ -52,7 +55,7 @@ def test_label_recursion(monkeypatch):
     # Stop plt.show() blocking tests
     global drawing_flag
     drawing_flag = False
-    monkeypatch.setattr(plt, "show", set_drawing_flag)
+    monkeypatch.setattr(matplotlib.pyplot, "show", set_drawing_flag)
 
     # add your geometry template here using PyMotorCAD
     r1 = create_triangle_reg(Coordinate(0, 0))
@@ -63,7 +66,214 @@ def test_label_recursion(monkeypatch):
 
     monkeypatch.setattr(ansys.motorcad.core.geometry_drawing, "_MAX_RECURSION", 1)
     with pytest.warns():
-        draw_objects([r1, r2, r3])
+        draw_objects([r1, r2, r3], label_regions=True)
+
+
+def test_drawing_base(mc):
+    # Test that various parameters generate images without causing fatal errors.
+    gt = mc.get_geometry_tree()
+    with tempfile.TemporaryDirectory() as temp_dir:
+        path = os.path.join(temp_dir, "base_tree.png")
+        draw_objects(
+            gt, save=path, axes=False, toggle_regions="Shaft", title="GT", draw_points=True
+        )
+
+
+def test_drawing_full_symmetry(mc):
+    # Test that various parameters generate images without causing fatal errors.
+    gt = mc.get_geometry_tree()
+    with tempfile.TemporaryDirectory() as temp_dir:
+        path = os.path.join(temp_dir, "full_symmetry.png")
+        draw_objects(
+            gt,
+            save=path,
+            optimise=True,
+            full_geometry=True,
+        )
+
+
+def test_drawing_region_points_labels(mc):
+    # Test that various parameters generate images without causing fatal errors.
+    gt = mc.get_geometry_tree()
+    with tempfile.TemporaryDirectory() as temp_dir:
+        path = os.path.join(temp_dir, "rotor.png")
+        draw_objects(
+            gt["Rotor"],
+            save=path,
+            label_regions=True,
+            draw_points=True,
+            dpi=800,
+        )
+
+
+def test_drawing_entities_points(mc):
+    # Test that various parameters generate images without causing fatal errors.
+    gt = mc.get_geometry_tree()
+    with tempfile.TemporaryDirectory() as temp_dir:
+        path = os.path.join(temp_dir, "rotor_entities.png")
+        draw_objects(
+            gt["Rotor"].entities,
+            save=path,
+            draw_points=True,
+        )
+
+
+def test_drawing_region_list(mc):
+    # Test that various parameters generate images without causing fatal errors.
+    gt = mc.get_geometry_tree()
+    with tempfile.TemporaryDirectory() as temp_dir:
+        path = os.path.join(temp_dir, "region_list.png")
+        draw_objects(
+            [gt["Rotor"], gt["Stator"]],
+            save=path,
+        )
+
+
+def test_scroll1(mc):
+    gt = mc.get_geometry_tree()
+    with tempfile.TemporaryDirectory() as temp_dir:
+        path = os.path.join(temp_dir, "scroll_test.png")
+        region_drawing = draw_objects(gt, expose_region_drawing=True, save=path)
+        first_top = region_drawing.check.labels[0].get_text()
+        next_top = region_drawing.check.labels[1].get_text()
+        # Check nothing occurs if at top
+        region_drawing.cycle_check("up")
+        region_drawing.cycle_check("down")
+        assert region_drawing.check.labels[0].get_text() == next_top
+
+        region_drawing.cycle_check("up")
+        assert region_drawing.check.labels[0].get_text() == first_top
+
+
+def test_scroll2(mc):
+    gt = mc.get_geometry_tree()
+    with tempfile.TemporaryDirectory() as temp_dir:
+        path = os.path.join(temp_dir, "scroll_test.png")
+        region_drawing = draw_objects(gt, expose_region_drawing=True, save=path)
+        bottom = list(region_drawing.keys_and_labels.forward.values())[-1]
+        for i in range(0, len(list(region_drawing.keys_and_labels.forward))):
+            region_drawing.cycle_check("down")
+
+        # Check it hasn't scrolled more than the maximum
+        assert region_drawing.check.labels[-1].get_text() == bottom
+
+
+def test_scroll3(mc):
+    # Test scroll does nothing if there are not enough items
+    gt = mc.get_geometry_tree()
+    with tempfile.TemporaryDirectory() as temp_dir:
+        path = os.path.join(temp_dir, "scroll_test.png")
+        region_drawing = draw_objects(
+            gt.get_subtree("Stator"), expose_region_drawing=True, save=path
+        )
+        first_top = region_drawing.check.labels[0].get_text()
+        region_drawing.cycle_check("down")
+        assert region_drawing.check.labels[0].get_text() == first_top
+
+
+def test_check(mc):
+    gt = mc.get_geometry_tree()
+    with tempfile.TemporaryDirectory() as temp_dir:
+        path = os.path.join(temp_dir, "check_test.png")
+        region_drawing = draw_objects(gt, expose_region_drawing=True, save=path)
+        current_housing = region_drawing.object_states["Housing"]
+        region_drawing.func("Housing")
+        assert current_housing != region_drawing.object_states["Housing"]
+
+
+def test_draw_entity(mc):
+    gt = mc.get_geometry_tree()
+    with tempfile.TemporaryDirectory() as temp_dir:
+        path = os.path.join(temp_dir, "entity.png")
+        draw_objects(gt["Rotor"].entities[0], save=path)
+
+
+def test_draw_coordinate(mc):
+    gt = mc.get_geometry_tree()
+    with tempfile.TemporaryDirectory() as temp_dir:
+        path = os.path.join(temp_dir, "coordinate.png")
+        draw_objects(gt["Rotor"].points[0], save=path)
+
+
+def test_draw_list_of_coordinates(mc):
+    gt = mc.get_geometry_tree()
+    with tempfile.TemporaryDirectory() as temp_dir:
+        path = os.path.join(temp_dir, "coordinate_list.png")
+        # path = r"C:\Workspace\pymotorcad\tests\test_files\temp_files\coordinate_list.png"
+        draw_objects(gt["Rotor"].points, save=path)
+
+
+def test_draw_list_of_coordinate_and_entity(mc):
+    gt = mc.get_geometry_tree()
+    with tempfile.TemporaryDirectory() as temp_dir:
+        path = os.path.join(temp_dir, "coordinate_entity_list.png")
+        draw_objects([gt["Rotor"].points[0], gt["Rotor"].entities[1]], save=path)
+
+
+def test_draw_list_of_coordinates_and_entities(mc):
+    gt = mc.get_geometry_tree()
+    with tempfile.TemporaryDirectory() as temp_dir:
+        path = os.path.join(temp_dir, "coordinate_entity_list.png")
+        to_draw = []
+        to_draw.extend(gt["Rotor"].points)
+        to_draw.extend(gt["Rotor"].entities)
+        draw_objects(to_draw, save=path)
+
+
+def test_draw_list_of_coordinate_and_region(mc):
+    gt = mc.get_geometry_tree()
+    with tempfile.TemporaryDirectory() as temp_dir:
+        path = os.path.join(temp_dir, "coordinate_region.png")
+        # path = r"C:\Workspace\pymotorcad\tests\test_files\temp_files\coordinate_region_list.png"
+        draw_objects([gt["Rotor"].points[0], gt["Rotor"]], save=path)
+
+
+def test_draw_list_of_coordinate_and_region_and_entity(mc):
+    gt = mc.get_geometry_tree()
+    with tempfile.TemporaryDirectory() as temp_dir:
+        path = os.path.join(temp_dir, "coordinate_entity_region.png")
+        draw_objects([gt["Rotor"].points[0], gt["Rotor"], gt["Shaft"].entities[2]], save=path)
+
+
+def test_draw_list_of_region_and_entity(mc):
+    gt = mc.get_geometry_tree()
+    with tempfile.TemporaryDirectory() as temp_dir:
+        path = os.path.join(temp_dir, "entity_region.png")
+        draw_objects([gt["Rotor"], gt["Shaft"].entities[2]], save=path)
+
+
+def test_draw_list_of_coordinates_and_regions_and_entities(mc):
+    gt = mc.get_geometry_tree()
+    with tempfile.TemporaryDirectory() as temp_dir:
+        path = os.path.join(temp_dir, "coordinates_entities_regions_list.png")
+        to_draw = []
+        to_draw.extend(gt["Rotor"].points)
+        to_draw.extend(gt["Shaft"].entities)
+        to_draw.append(gt["Rotor"])
+        to_draw.append(gt["RotorDuctFluidRegion"])
+        to_draw.append(gt["RotorDuctFluidRegion_1"])
+        draw_objects(to_draw, save=path)
+
+
+def test_bidict():
+    test_dict = BiDict()
+    test_dict.insert("key1", "value1")
+    test_dict.insert("key1", "value2")
+    assert test_dict.get_forward("key1") == "value2"
+    assert test_dict.get_backward("value2") == "key1"
+    test_dict.remove_by_value("value2")
+    assert test_dict.forward == dict()
+    assert test_dict.backward == dict()
+
+    test_dict = BiDict()
+    test_dict.insert("key1", "value1")
+    test_dict.insert("key2", "value1")
+    assert test_dict.get_forward("key2") == "value1"
+    assert test_dict.get_backward("value1") == "key2"
+
+    test_dict.remove_by_key("key2")
+    assert test_dict.forward == dict()
+    assert test_dict.backward == dict()
 
 
 # def test_draw_objects_debug(mc, monkeypatch):
