@@ -1,4 +1,4 @@
-# Copyright (C) 2022 - 2024 ANSYS, Inc. and/or its affiliates.
+# Copyright (C) 2022 - 2026 ANSYS, Inc. and/or its affiliates.
 # SPDX-License-Identifier: MIT
 #
 #
@@ -32,7 +32,7 @@ import pytest
 import ansys.motorcad.core as pym
 import ansys.motorcad.core as pymotorcad
 from ansys.motorcad.core import MotorCAD, MotorCADError, MotorCADWarning
-from ansys.motorcad.core.rpc_client_core import _MotorCADConnection
+from ansys.motorcad.core.rpc_client_core import MOTORCAD_EXE_GLOBAL, _MotorCADConnection
 
 
 def test__find_free_motor_cad(mc):
@@ -56,6 +56,7 @@ def test_set_server_ip():
 
 
 def test_set_motorcad_exe():
+    save_global_exe = MOTORCAD_EXE_GLOBAL
     test_path = r"test_path/test"
 
     pymotorcad.set_motorcad_exe(test_path)
@@ -63,7 +64,7 @@ def test_set_motorcad_exe():
     assert pymotorcad.rpc_client_core._find_motor_cad_exe() == test_path
 
     # reset path
-    pymotorcad.set_motorcad_exe("")
+    pymotorcad.set_motorcad_exe(save_global_exe)
     pymotorcad.rpc_client_core._find_motor_cad_exe()
 
 
@@ -336,3 +337,21 @@ def test__resolve_localhost():
         raise e
     finally:
         mc2.quit()
+
+
+def test_blackbox_licencing():
+    mc1 = MotorCAD(use_blackbox_licence=True)
+    # Not sure it's possible to assert that only a blackbox licence was consumed
+    # Just check it works for now
+    mc1.get_licence()
+
+    mc2 = MotorCAD(use_blackbox_licence=False)
+    # Not sure it's possible to assert that only a non-blackbox licence was consumed
+    # Just check it works for now
+    mc2.get_licence()
+
+    mc3 = MotorCAD()
+    # Not sure it's possible to check which licence type has been used, and whether this
+    # matches the default setting
+    # Just check it works for now
+    mc3.get_licence()
